### 2021-06-29 #3433 @arfon
In the meantime, changing the paper header to quote the affiliations for Florian Wiesinger should fix your paper compilation:

```
  - name: Florian Wiesinger
    orcid: 0000-0002-5597-6057
    affiliation: "1,2"
```
