### 2021-06-30 3551a16
Quote affiliation
