### 2021-07-05 feff431
Fix apodization in TGV
