### 2021-07-05 aa76846
Minor tweak to lo spokes gridding
