### 2021-07-06 #3433 @spinicist
Hello,

From the recon side you could try jtamir, martink84, uecker, or any of the other contributors to BART (https://github.com/mrirecon/bart/graphs/contributors) or SigPy (https://github.com/mikgroup/sigpy/graphs/contributors).

You could also try agahkarakuzu from qMRLab (https://github.com/mikgroup/sigpy/graphs/contributors)

Thanks
