### 2021-07-06 #3433 @MartinK84
I would be available as a reviewer and have experience in the relevant field (working with 3D non-cartesian MRI data and have used comparable toolboxes such as BART). I have to warn, however, that I'm very busy at the moment and will probably need several weeks to be able to finish reviewing this contribution thoroughly.
