### 2021-07-06 #3433 @emdupre
Hi @spinicist, thanks for your submission to JOSS ! As @arfon mentioned above, we've swapped handling editors, here, so I'll be acting as editor for your submission.

@MartinK84, thank you for your willingness to review ! I'll wait a bit longer to see if other reviewers might have more availability, but I will of course keep you up to date.
