### 2021-07-09 #3433 @emdupre
Thanks for your willingness to review, @uecker  ! You can find more information on the JOSS review process on [the JOSS documentation](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), including the [checklist we ask you to complete](https://joss.readthedocs.io/en/latest/review_checklist.html) during a review.

If that all sounds alright to you, I can add you as a reviewer on this submission.
