### 2021-07-12 #3433 @emdupre
Hi @spinicist ! While we're waiting to confirm the reviewers, a few quick notes from me:

- The abbreviation RIESLING is not defined in the paper, although I assume it stands for your title "Radial Interstices Enable Speedy Low-volume Imaging." Please clarify or otherwise make this explicit in the text.
- Your paper doesn't yet have a [Statement of Need section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), which will help to explain why this software is useful to the field.
- The current Summary section lacks a broader introduction for the non-specialist reader. Important terms like Zero Echo Time are considered but not defined.

Please check the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for an idea of the level of detail that we expect here !
