### 2021-07-12 9927b8f
Add option to phantom to use external SENSE maps
