+++ src/io_hd5.cpp
@@ -4,16 +4,18 @@
 #include <filesystem>
 #include <fmt/format.h>
 
-namespace HD5 {
+namespace Keys {
+std::string const Info = "info";
+std::string const Meta = "meta";
+std::string const Noncartesian = "noncartesian";
+std::string const Cartesian = "cartesian";
+std::string const Image = "image";
+std::string const Trajectory = "trajectory";
+std::string const SDC = "sdc";
+std::string const SENSE = "sense";
+} // namespace Keys
 
-std::string const KeyInfo = "info";
-std::string const KeyMeta = "meta";
-std::string const KeyNoncartesian = "noncartesian";
-std::string const KeyCartesian = "cartesian";
-std::string const KeyImage = "image";
-std::string const KeyTrajectory = "trajectory";
-std::string const KeySDC = "sdc";
-std::string const KeySENSE = "sense";
+namespace HD5 {
 
 void Init(Log &log)
 {
@@ -86,7 +88,7 @@ void Writer::writeInfo(Info const &info)
   hsize_t dims[1] = {1};
   auto const space = H5Screate_simple(1, dims, NULL);
   hid_t const dset =
-      H5Dcreate(handle_, KeyInfo.c_str(), info_id, space, H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT);
+      H5Dcreate(handle_, Keys::Info.c_str(), info_id, space, H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT);
   if (dset < 0) {
     log_.fail("Could not create info struct, code: {}", dset);
   }
@@ -102,7 +104,7 @@ void Writer::writeInfo(Info const &info)
 void Writer::writeMeta(std::map<std::string, float> const &meta)
 {
   log_.info("Writing meta data");
-  auto m_group = H5Gcreate(handle_, KeyMeta.c_str(), H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT);
+  auto m_group = H5Gcreate(handle_, Keys::Meta.c_str(), H5P_DEFAULT, H5P_DEFAULT, H5P_DEFAULT);
 
   hsize_t dims[1] = {1};
   auto const space = H5Screate_simple(1, dims, NULL);
@@ -122,37 +124,37 @@ void Writer::writeMeta(std::map<std::string, float> const &meta)
 void Writer::writeTrajectory(Trajectory const &t)
 {
   writeInfo(t.info());
-  HD5::store_tensor(handle_, KeyTrajectory, t.points(), log_);
+  HD5::store_tensor(handle_, Keys::Trajectory, t.points(), log_);
 }
 
 void Writer::writeSDC(R2 const &sdc)
 {
-  HD5::store_tensor(handle_, KeySDC, sdc, log_);
+  HD5::store_tensor(handle_, Keys::SDC, sdc, log_);
 }
 
 void Writer::writeSENSE(Cx4 const &s)
 {
-  HD5::store_tensor(handle_, KeySENSE, s, log_);
+  HD5::store_tensor(handle_, Keys::SENSE, s, log_);
 }
 
 void Writer::writeNoncartesian(Cx4 const &t)
 {
-  HD5::store_tensor(handle_, KeyNoncartesian, t, log_);
+  HD5::store_tensor(handle_, Keys::Noncartesian, t, log_);
 }
 
 void Writer::writeCartesian(Cx4 const &t)
 {
-  HD5::store_tensor(handle_, KeyCartesian, t, log_);
+  HD5::store_tensor(handle_, Keys::Cartesian, t, log_);
 }
 
 void Writer::writeImage(R4 const &t)
 {
[…]
