### 2021-07-12 d69d674
Tweaked paper
