+++ docs/paper.md
@@ -26,9 +26,22 @@ bibliography: paper.bib
 
 # Summary
 
-- Zero Echo Time (ZTE) imaging is a unique form of MRI with many interesting benefits, such as near silent operation [@Ljungberg], but also technical challenges such as dead-time gap artefacts [@Froidevaux]. The 3D radial trajectory inherent to ZTE can be challenging to reconstruct due to large memory requirements and convergence problems from uneven k-space sampling [@Zwart]. Existing reconstruction toolboxes such as BART [@BART] or SigPy provide high-quality implementations of many algorithms but have not been tuned for the specific case of 3D non-cartesian trajectories, which can lead to suboptimal performance.
-- The RIESLING toolbox (http://github.com/spinicist/riesling) has been written and tuned for the specific application of 3D radial ZTE imaging, but can be used to reconstruct other non-cartesian trajectories such as spirals as well. RIESLING is written with a modern C++ toolchain and utilizes Eigen [@Eigen] for all core operations, providing high performance. Data is stored in the HDF5 format allowing easy interoperation with other tools.
-- RIESLING provides multiple reconstruction strategies, including classic conjugate-gradient SENSE [@cgSENSE] and Total-Generalized Variation [@TGV].
-- We are actively using RIESLING for our own studies and hope that it will be of interest to other groups using such sequences.
+- Standard MRI methods acquire Fourier-encoded data on a regularly spaced Cartesian grid. Noncartesian MRI methods abandon this grid and instead acquire data on arbitrary trajectories in k-space, which can lead to advantages such as motion robustness and more flexible pulse sequence design.
+- Zero Echo Time (ZTE) imaging is a specialised form of MRI with an inherently noncartesian three-dimensional radial trajectory [@Ljungberg]. ZTE imaging has many interesting benefits, such as near silent operation, but also technical challenges such as dead-time gap artefacts [@Froidevaux].
+- We have developed a toolbox, named Radial Interstices Enable Speedy Low-volume imagING (RIESLING), tuned for high performance reconstruction of MRI data acquired with 3D noncartesian k-space trajectories. While our group has a focus on ZTE imaging, RIESLING is suitable for all 3D noncartesian trajectories.
+
+# Statement of Need
+
+3D noncartesian trajectories can be challenging to reconstruct compared to other MRI trajectories. The major issues are:
+- Inseparability of the trajectory dimensions. In Cartesian imaging it is often possible to consider one of the image dimensions separately to the others, leading to algorithmic complexity reduction and memory savings. 3D noncartesian must consider all dimensions simultaneously.
+- Oversampling requirements. In order to obtain good image quality, it is necessary to oversample the reconstruction grid. Coupled with the inseparability of the 3D problem, this leads to large memory requirements, which can often exceed that available on a GPU.
+- Sample density compensation. Noncartesian trajectories can lead to uneven k-space sampling, the correction of which is equivalent to preconditioning a linear system [@cgSENSE]. For 2D noncartesian trajectories, it is generally not necessary to correct for this in the context of an iterative reconstruction, as the problem will still converge. For 3D noncartesian trajectories, such preconditioning is essential to ensure reasonable convergence properties.
+
+Existing MRI reconstruction toolboxes such as BART [@BART] or SigPy provide high-quality implementations of many reconstruction algorithms but have prioritised applications other than 3D noncartesian reconstruction. We hence created a dedicated toolbox for this, including specific features such as:
+- A configurable grid oversampling factor
+- A thread-safe gridding implementation suitable for multi-core CPUs
+- Integrated sample density compensation functions [@Zwart]
+
[…]
