+++ docs/paper.bib
@@ -54,6 +54,17 @@
     year = {2015}
 }
 
+@inproceedings{SigPy,
+address = {Montreal, QC, Canada},
+author = {Ong, Frank and Lustig, Michael},
+booktitle = {Proc. Intl. Soc. Mag. Reson. Med. 27},
+file = {:Users/emilljungberg/Documents/Mendeley Library/Ong, Lustig - 2019.pdf:pdf},
+pages = {4819},
+title = {{SigPy: A Python Package for High Performance Iterative Reconstruction}},
+year = {2019}
+}
+
+
 @MISC{Eigen,
   author = {{Guennebaud, Ga\"{e}l}, {Jacob, Beno\^{i}t}},
   title = {Eigen v3},
