### 2021-07-12 #3433 @spinicist
Hello,

We have expanded the paper and included the acronym definition. The statement of need has been split out of the summary and clarified. Echo Time (TE) is a general MRI term, and the cited paper for ZTE is a comprehensive review of how it may be reduced to zero.
