### 2021-07-14 #3433 @neurolabusc
@emdupre I feel this work is outside my expertise. I do think reviewers like Uecker from the BART project would be ideal. You might also want to consider the developers of [mripy](https://github.com/peng-cao/mripy).
