### 2021-07-15 #3433 @emdupre
Thank you, @uecker ! I'll add you as a reviewer now.

Thanks, too, @neurolabusc for your quick response and for sharing mripy !

From an offline discussion, I'm going to avoid assigning you as a reviewer here, @agahkarakuzu, so you're not over-committed with concurrent JOSS reviews.

@MartinK84, would you still be willing to review ? I know that you said you might need a several week delay -- would you be able to give a more concrete estimate of that timeline (e.g. one month) ? I appreciate any additional detail you could offer, here !
