### 2021-07-15 #3433 @emdupre
Thank you, @MartinK84 !

As JOSS is still in [reduced service mode](https://blog.joss.theoj.org/#reduced-service-mode), a month would be a very reasonable timeline, here.  I'll go ahead and add you as a reviewer to the submission now !
