### 2021-07-27 c1bbc79
Set required CMake to 3.10
