### 2021-07-27 a2c8eb4
Ensure build directory exists
