### 2021-07-27 e2ae824
Add check for ninja to bootstrap.sh
