### 2021-07-28 7727cf0
Fix ambiguous overload on GCC 8
