### 2021-07-29 spinicist/riesling#22 @MartinK84
Purpose of read_gap
I'm currently in the process of trying to reconstruct my own 3D-UTE data with riesling. Creating the relevant k-space and trajectory shapes is okay in my opinion and well enough documented (after finding the tutorial notebook in the other repository). However, I'm struggling a bit with the info header because not all parameters are documented (looking here: https://riesling.readthedocs.io/en/latest/docs/data.html).

Especially the `read_gap` parameter is a bit puzzling to me. I would assume that this is the number of additional data points that should be discarded at the beginning of each readout/ADC. But I'm not sure and would need some clarification and would also suggest updating this info in the documentation.

I think most of the other info parameters are pretty self-explanatory, however, it would still be great if more details could be added to the documentation. Like the unit of TR, etc.
