### 2021-07-29 spinicist/riesling#22 @MartinK84
I'm also currently not fully sure about the `spokes_hi` parameter, is this supposed to be the number of acquired spokes or the Nyquist number of spokes for a fully sampled k-space that corresponds to the given matrix size? I'm asking because for the Phantom dataset from the tutorial notebook the number of spokes in k-space is 24575 but the info.spokes_hi is 21845, so the numbers do not match :astonished:
