### 2021-07-29 spinicist/riesling#22 @spinicist
> Especially the `read_gap` parameter is a bit puzzling to me. I would assume that this is the number of additional data points that should be discarded at the beginning of each readout/ADC. But I'm not sure and would need some clarification and would also suggest updating this info in the documentation.

`read_gap` is indeed the number of points that fall into the ZTE dead-time gap and should be discarded. I will update the documentation, apologies that this was missed.

> I think most of the other info parameters are pretty self-explanatory, however, it would still be great if more details could be added to the documentation. Like the unit of TR, etc.

Yes, this will also be added. Technically the units of TR and the image dimensions are only used when constructing a NIFTI header, and hence should be seconds and millimeters respectively.
