### 2021-07-29 spinicist/riesling#22 @spinicist
> I'm also currently not fully sure about the `spokes_hi` parameter, is this supposed to be the number of acquired spokes or the Nyquist number of spokes for a fully sampled k-space that corresponds to the given matrix size? I'm asking because for the Phantom dataset from the tutorial notebook the number of spokes in k-space is 24575 but the info.spokes_hi is 21845, so the numbers do not match 😲

The use of `spokes_hi` and `spokes_lo` is one place where riesling is tailored towards ZTE acquisitions, specifically those with a WASPI-type acquisition with a second low-resolution k-space to fill the dead-time gap.

From the documentation: "`spokes_hi` sets the number of spokes in the main, high-resolution k-space acquisition." For UTE data this should be the number of spokes in the acquisition, and `spokes_lo` should be zero.

The tutorial includes the use of a low-resolution k-space, hence the discrepancy between `spokes_hi` and the total number of spokes. The total number should be equal to `spokes_hi` + `spokes_lo`.

I am open to suggestions as to renaming these header fields for clarity.
