### 2021-07-29 spinicist/riesling#22 @MartinK84
I got confused because I was thinking if `spokes_lo` could mean something like fully sampled center-region of k-space for some kind of k-zero based calibration.  With what you wrote I think the names `spokes_hi`+ `spokes_lo` are fine and make sense, it should just be clarified in the documentation that both are tailored for ZTE acquisitions. Maybe just explicitly state this and also state how they are supposed to be set for 3D-UTE.
