### 2021-07-29 spinicist/riesling#23 @MartinK84
Error using h5 created with MATLAB
I have been trying to use my own data for the reconstruction of 3D-UTE data based on the tutorial notebook. What I did so far in MATLAB was reshape my data structures to the format required by riesling and create a header info struct. However, when trying a simple RSS reco riesling fails directly after or during loading of the trajectory.

```
./riesling recon --verbosity=3 --rss --sdc=pipe UTE.h5
Starting operation: recon
Read wisdom successfully
Opened file UTE.h5 for reading
Reading info
Reading trajectory
terminate called after throwing an instance of 'std::bad_alloc'
  what():  std::bad_alloc
Aborted (core dumped)
```

The header I think and hope is set correctly:

```
./riesling hdr UTE.h5
Header
Matrix: 96 96 96
Voxel-size: 2.08333 2.08333    1.25
Read points: 122 Gap: 0
Hi-res spokes: 8002 Lo-res spokes: 0 Lo-res scale: 1
Channels: 16 Volumes: 1 TR: 1
Origin: 0 0 0
Direction:
1 0 0
0 1 0
0 0 1
```

The relevant MATLAB output used for creating the h5 file:

```
>> size(kspace_UTE)
ans =
          16         122        8002

>> size(traj_UTE)
ans =
           3         122        8002

>> write_riesling('UTE.h5', single(kspace_UTE), single(traj_UTE), info_UTE);
```

at this point, I'm a bit out of ideas without going into the code and starting to add additional debug output, any hints or ideas what I could be missing?
