### 2021-07-29 spinicist/riesling#23 @spinicist
To date I've only seen `std::bad_alloc` thrown when riesling has run out of memory - either due to a very large reconstruction, or on a grid job with memory constraints. Given the dimensions of your problem this seems unlikely.

Is it possible for you to share the `UTE.h5` file so I can investigate?
