### 2021-07-29 spinicist/riesling#23 @MartinK84
I will send you a link to the data via Twitter in a minute (don't want to publicly share it). Memory should not be an issue as that computation server has 768GB of RAM.
