### 2021-07-29 c0f7ce4
Switch Linux CI to non-AVX2
