### 2021-07-29 spinicist/riesling#23 @spinicist
The issue was thankfully a simple one - in your `Info` struct you had put `spokes_low` instead of `spokes_lo`. It appears the HDF5 library will happily skip over missing/mis-spelled fields and leave junk values there. On my machine `spokes_lo` was consequently being filled with a very large number, if the same happened on your machine I suspect it could overflow even 768 GB!

I have now added code to check that all the header fields riesling expects are actually present, and report if any are not found. I also changed the name of the header field in your file and managed to run the reconstruction.
