### 2021-07-29 spinicist/riesling#23 @MartinK84
Awesome thanks, will soon give it another try.

Does it make sense to give the MATLAB wrapper some additional features? It could basically also check the header names and possibly data types to avoid any fallout in the C++ backend.
