### 2021-07-29 26f23a1
Fix bug with header field checking
