### 2021-07-29 d87cb4c
Modified read_riesling() to also load image data
