### 2021-07-29 906272b
Modified read_riesling() to also load image data
