### 2021-07-29 3fae795
Update scn::scan to v0.4 version
