### 2021-07-30 53df64b
Add header info checks to Matlab code
