### 2021-07-30 7077fbd
Improve documentation of the data format.
