### 2021-07-30 spinicist/riesling#23 @spinicist
I have added checking of the header fields to the MATLAB code.

Note that I renamed the MATLAB files for consistency - they now all start with `riesling_`.
