### 2021-07-30 spinicist/riesling#22 @spinicist
I have updated the documentation, I hope it is clearer now.
