### 2021-08-02 405316b
Some additions to usage of RIESLING and added Ninja dependency
