### 2021-08-02 spinicist/riesling#26 @MartinK84
Unit Test
RIESLING appears to have a Test folder with some automated Unit Tests. I assume that those are run automatically by GitHub?
Is there any way to run the Tests manually from the local command line to check if changes made break anything or just to confirm full functionality? Both the README and CONTRIB documentations unfortunately do not specify anything on this topic.
