### 2021-08-02 5431f01
Added notes on tests/code
