+++ docs/contents.rst
@@ -8,3 +8,4 @@ Contents
    recon
    util
    data
+   devel
