### 2021-08-02 spinicist/riesling#26 @spinicist
Yes, there are some unit tests. They are not currently comprehensive but do cover the core gridding/FFT functions. For full functionality, I regard the tutorial notebook as a full Acceptance Test. If that runs, everything should be working.

These tests were previously built by the `prufung` target, but I have pushed an update that changes this to `riesling-tests` for clarity. I have added some notes on these to a new page in the documentation https://riesling.readthedocs.io/en/latest/docs/devel.html
