### 2021-08-02 spinicist/riesling#26 @spinicist
PS Yes, these tests are run as part of the Github CI
