### 2021-08-02 3deb01e
Adjusted dependencies, help guideline and fixed some more typos
