### 2021-08-02 928f21a
Some additions to usage of RIESLING and added Ninja dependency (#25)
* Some additions to usage of RIESLING, help guideline and fixed some more typos
