+++ docs/recon.rst
@@ -18,7 +18,7 @@ Non-cartesian MRI data requires a Non-Uniform FFT (NUFFT) instead of a simple FF
 
 The final gridding option is ``--fast-grid``. RIESLING uses a parallelized gridding operation. In order to avoid multiple threads writing to the same Cartesian k-space point, the trajectory is sorted by Cartesian k-space location and threads each process their own chunk of the sorted co-ordinates. However, there may still be race conditions at the edge of each chunk, particularly for small images with highly oversampled k-space centers. To prevent this, each thread uses it's own workspace, and then these are combined into the final grid at the end. This process is thread-safe but doubles the memory requirements for the gridding operations. The ``--fast-grid`` option makes the threads write directly into the final grid, reducing memory consumption but at the risk of some writes into the grid being overwritten. When gridding high resolution images on a small number of threads, e.g. 10 or fewer, the probability of a race condition is vanishingly small. Use at your own risk.
 
-The gridding step can also compensate for the increased density of samples in the oversampled central k-space region with most non-Cartesian trajectories. This step is often ommitted in 2D non-Cartesian iterative reconstructions, but is essential for reasonable convergence in 3D non-Cartesian reconstruction. The ``--sdc`` option controls the sample density compensation method - valid values are "none" to turn it off, "radial" for analytic radial densities, and "pipe" for the iterative method of Pipe et al. The default is "pipe". It is also possible to pre-calculate the densities of a given trajectory using the ``riesling sdc`` command and then pass in the resulting file, i.e. ``--sdc=file-sdc.h5``. The weighting of the compensation can be reduced using the ``--sdc_exp`` option - the exponent is applied to all k-space densitites equally and values between 0 and 1 make sense.
+The gridding step can also compensate for the increased density of samples in the oversampled central k-space region with most non-Cartesian trajectories. This step is often omitted in 2D non-Cartesian iterative reconstructions, but is essential for reasonable convergence in 3D non-Cartesian reconstruction. The ``--sdc`` option controls the sample density compensation method - valid values are "none" to turn it off, "radial" for analytic radial densities, and "pipe" for the iterative method of Pipe et al. The default is "pipe". It is also possible to pre-calculate the densities of a given trajectory using the ``riesling sdc`` command and then pass in the resulting file, i.e. ``--sdc=file-sdc.h5``. The weighting of the compensation can be reduced using the ``--sdc_exp`` option - the exponent is applied to all k-space densities equally and values between 0 and 1 make sense.
 
 Due to the oversampled central region in most non-Cartesian trajectories reasonable channel sensitivities can be extracted directly from the data. This is the default option. Tikhonov regularization can be applied to the sensitivities using the ``--lambda`` option (the value should be approximately the same as the background noise in the ``--rss`` reconstruction). In a multi-volume reconstruction, the sensitivities are taken from the last volume by default but can be specified using ``--senseVolume``, or can be taken from an external file using ``--sense``.
 
