### 2021-08-03 spinicist/riesling#26 @MartinK84
Thanks for the devel docs update, that's quite important! `riesling-tests` also now works from the Commandline -> issue closed
