### 2021-08-03 #3500 @MartinK84
@emdupre I have completed my review

In my opinion, the submission meets all requirements for publication in JOSS and I highly recommend acceptance. I can confirm that all claims made by the authors are reasonable and valid. Without much work, I was able to use and apply RIESLING to my own data with very good results. There were some minor issues regarding the installation, documentation, and usage which I have resolved together with the corresponding author @spinicist over in their repository.
