### 2021-08-04 #3500 @emdupre
Thank you for your review, @MartinK84 !

I am cross-linking that these discussions have taken place in https://github.com/spinicist/riesling/issues/22, https://github.com/spinicist/riesling/issues/23, and https://github.com/spinicist/riesling/issues/26. If I have missed any additional references, please don't hesitate to add them here.
