### 2021-08-04 #3500 @emdupre
👋 Hi @uecker -- have you had a chance to take a look at the paper, and start working through the review checklist ? If you post issues on the software repo, please add a pointer to this issue so they are cross-linked.
