### 2021-08-20 dcec72c
Started making some Operators
