+++ bench/main.cpp
@@ -1,0 +1,4 @@
+#define CATCH_CONFIG_ENABLE_BENCHMARKING
+#define CATCH_CONFIG_MAIN
+
+#include <catch2/catch.hpp>
