### 2021-09-01 74d93d4
Tweaked main code with benchmark results
