### 2021-09-01 271b6b2
Put tests and benchmarks behind options
