### 2021-09-05 spinicist/riesling#27 @uecker
Debian 10 linking fails.
Hi Toby,
on my Debian buster notebook linking fails (after two hours compiling or so). Any ideas?

[100%] Linking CXX executable riesling
/usr/bin/ld: CMakeFiles/riesling.dir/src/main_phantom.cpp.o: in function `main_phantom(args::Subparser&)':
main_phantom.cpp:(.text+0x35da): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: main_phantom.cpp:(.text+0x3619): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: main_phantom.cpp:(.text+0x3628): undefined reference to `std::filesystem::__cxx11::path::replace_extension(std::filesystem::__cxx11::path const&)'
/usr/bin/ld: libvineyard.a(fft_util.cpp.o): in function `FFT::WisdomPath[abi:cxx11]()':
fft_util.cpp:(.text+0xd2e): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: fft_util.cpp:(.text+0xda5): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: fft_util.cpp:(.text+0xeb0): undefined reference to `std::filesystem::__cxx11::path::has_root_directory() const'
/usr/bin/ld: fft_util.cpp:(.text+0xffc): undefined reference to `std::filesystem::__cxx11::path::has_filename() const'
/usr/bin/ld: fft_util.cpp:(.text+0x106b): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: fft_util.cpp:(.text+0x1161): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: libvineyard.a(io_hd5.cpp.o): in function `HD5::Reader::Reader(std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, Log&)':
io_hd5.cpp:(.text+0xd9a): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: io_hd5.cpp:(.text+0xda2): undefined reference to `std::filesystem::status(std::filesystem::__cxx11::path const&)'
/usr/bin/ld: libvineyard.a(parse_args.cpp.o): in function `OutName(std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&)':
parse_args.cpp:(.text+0x36a): undefined reference to `std::filesystem::__cxx11::path::_M_split_cmpts()'
/usr/bin/ld: parse_args.cpp:(.text+0x42b): undefined reference to `std::filesystem::__cxx11::path::replace_extension(std::filesystem::__cxx11::path const&)'
collect2: error: ld returned 1 exit status
make[2]: *** [CMakeFiles/riesling.dir/build.make:345: riesling] Fehler 1
make[1]: *** [CMakeFiles/Makefile2:73: CMakeFiles/riesling.dir/all] Fehler 2
make: *** [Makefile:130: all] Fehler 2
