### 2021-09-06 ad5cd78
Remove ITK as a dependency, use nifti directly
