+++ src/io_nifti.h
@@ -6,8 +6,3 @@
 template <typename T, int ND>
 extern void WriteNifti(
     Info const &info, Eigen::Tensor<T, ND> const &img, std::string const &fname, Log const &log);
-template <typename T>
-extern void WriteNifti(Eigen::Matrix<T, -1, -1> const &m, std::string const &fname, Log const &log);
-
-template <typename T, int ND>
-extern Eigen::Tensor<T, ND> ReadNifti(std::string const &fname, Log const &log);
