+++ src/io_nifti.cpp
@@ -1,112 +1,112 @@
 #include "io_nifti.h"
-#include "itkImage.h"
-#include "itkImageFileReader.h"
-#include "itkImageFileWriter.h"
-#include "itkImportImageFilter.h"
+#include "nifti1_io.h"
 #include <fmt/ostream.h>
+
+template <typename T>
+struct nifti_traits;
+
+template <>
+struct nifti_traits<float>
+{
+  constexpr static int dtype()
+  {
+    return NIFTI_TYPE_FLOAT32;
+  }
+  constexpr static int bytesper()
+  {
+    return 4;
+  }
+};
+
+template <>
+struct nifti_traits<std::complex<float>>
+{
+  constexpr static int dtype()
+  {
+    return NIFTI_TYPE_COMPLEX64;
+  }
+  constexpr static int bytesper()
+  {
+    return 8;
+  }
+};
 
 template <typename T, int ND>
 void WriteNifti(
     Info const &info, Eigen::Tensor<T, ND> const &img, std::string const &fname, Log const &log)
 {
-  using Image = itk::Image<T, ND>;
-  using Importer = itk::ImportImageFilter<T, ND>;
-  using Writer = itk::ImageFileWriter<Image>;
-
-  typename Importer::Pointer import = Importer::New();
-  typename Importer::SizeType sz;
-  typename Importer::IndexType st;
-  auto const dims = img.dimensions();
-  std::copy_n(dims.begin(), ND, &sz[0]);
-  std::fill_n(&st[0], ND, 0.);
-  typename Importer::RegionType const region{st, sz};
-  import->SetRegion(region);
-
-  double spacing[ND];
-  std::copy_n(info.voxel_size.begin(), 3, spacing);
+  nifti_image *ptr = nifti_simple_init_nim();
+  ptr->fname = nifti_makehdrname(fname.c_str(), NIFTI_FTYPE_NIFTI1_1, false, false);
+  ptr->iname = nifti_makeimgname(fname.c_str(), NIFTI_FTYPE_NIFTI1_1, false, false);
+  ptr->nvox = img.size();
+  ptr->xyz_units = NIFTI_UNITS_MM | NIFTI_UNITS_SEC;
+  ptr->dim[0] = ptr->ndim = 4;
+  ptr->dim[1] = ptr->nx = img.dimension(0);
+  ptr->dim[2] = ptr->ny = img.dimension(1);
+  ptr->dim[3] = ptr->nz = img.dimension(2);
   if constexpr (ND == 4) {
-    spacing[3] = info.tr;
+    ptr->dim[4] = ptr->nt = img.dimension(3);
+  } else {
+    ptr->dim[4] = ptr->nt = 1;
   }
-  import->SetSpacing(spacing);
+  ptr->dim[5] = ptr->nu = 1;
+  ptr->dim[6] = ptr->nv = 1;
+  ptr->dim[7] = ptr->nw = 1;
 
-  double origin[ND];
-  std::copy_n(info.origin.begin(), 3, origin);
-  if constexpr (ND == 4) {
-    origin[3] = 0;
-  }
-  import->SetOrigin(origin);
+  ptr->datatype = nifti_traits<T>::dtype();
+  ptr->nbyper = nifti_traits<T>::bytesper();
+  ptr->qform_code = NIFTI_XFORM_SCANNER_ANAT;
+  ptr->sform_code = NIFTI_XFORM_SCANNER_ANAT;
 
-  itk::Matrix<double, ND, ND> direction;
-  direction.Fill(0.);
-  // This appears to need a transpose. I don't know why
-  for (auto ir = 0; ir < 3; ir++) {
[…]
