### 2021-09-06 d8aca48
Adds workaround for GCC 8 std::filesystem
