### 2021-09-06 spinicist/riesling#27 @uecker
Am Montag, den 06.09.2021, 17:21 +0000 schrieb Tobias Wood:
> Hello,
>
> First a tangent - you correctly highlight that building the dependencies takes a long time. Most
> of that is spent compiling ITK and its dependencies, only to enable writing out nifti files. I
> have pushed a commit on main to remove ITK and use the nifti C library directly. I hope this
> should cut the first compile time by about half (FFTW is the other major time sink).
>
> Your actual problem appears to be a quirk of GCC 8 (which I think is the Debian 10 system
> compiler?) which requires an extra flag to link the C++ filesystem library correctly. There is a
> note at the bottom of the page here https://en.cppreference.com/w/cpp/filesystem about this that I
> previously did not know about.
>
> I have pushed a commit on a new branch `stdfs` (https://github.com/spinicist/riesling/tree/stdfs)
> which I hope has a work-around in `CMakeLists.txt`. Could you please try this branch and see what
> happens?

I installed libnifti-dev, but now run into:

CMake Error at cmake/vcpkg/scripts/buildsystems/vcpkg.cmake:782 (_find_package):
  By not providing "Findnifti.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "nifti", but
  CMake did not find one.

  Could not find a package configuration file provided by "nifti" with any of
  the following names:

    niftiConfig.cmake
    nifti-config.cmake

  Add the installation prefix of "nifti" to CMAKE_PREFIX_PATH or set
  "nifti_DIR" to a directory containing one of the above files.  If "nifti"
  provides a separate development package or SDK, be sure it has been
  installed.
Call Stack (most recent call first):
  CMakeLists.txt:14 (find_package)

-- Configuring incomplete, errors occurred!

Best,
Martin
