### 2021-09-06 dee6fd6
NIFTI should be uppercase in CMakeLists.txt
