### 2021-09-06 b418038
Incorrect variable name in CMakeLists
