### 2021-09-06 spinicist/riesling#27 @spinicist
Sadly I typed `nifti` instead of `NIFTI` as the package name in CMakeLists.txt, but because APFS is case-insensitive on my Mac CMake managed to find it anyway. But Linux is (correctly) case-sensitive, hence the failure which was picked up in the CI but I hadn't waited for that to finish.

This has now been fixed, please try the `stdfs` branch again. It built successfully in the Linux CI so fingers crossed.

(You should not need to install any libraries - vcpkg will build them from source)
