### 2021-09-07 spinicist/riesling#27 @uecker
The nifit problem is solved but not yet the original problem. (The pre-compiled binary works fine.)
