### 2021-09-07 spinicist/riesling#27 @spinicist
Okay. I will have to try and find a Linux system I can install the same version of GCC on and debug further.

Please can you confirm the exact version of GCC on your system?
