### 2021-09-07 spinicist/riesling#28 @uecker
References for Gridding
I wonder what gridding technique was implemented? Fessler's paper about min-max interpolation and a paper about nearest neighbor gridding are cited. But are these techniques really implemented?

For Toeplitz embedding I cite the ISMRM abstract by ETH and for gridding itself the papers by O’Sullivan and Jackson et al. (see also doc/references.txt in BART.)
