### 2021-09-07 #3500 @uecker
Two issues I raised so far:
https://github.com/spinicist/riesling/issues/27
https://github.com/spinicist/riesling/issues/28
