### 2021-09-07 spinicist/riesling#27 @uecker
The default system compiler is: gcc (Debian 8.3.0-6) 8.3.0

Unfortunately,  I never understood how cmake works, so I can not really help with this, but if you need other information let me know.
