### 2021-09-08 11ba196
Changed trajectory scaling
