### 2021-09-12 #3500 @uecker
**Conflict of interest**
- [X]  I confirm that I have read the JOSS conflict of interest (COI) policy and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
**Code of Conduct**
- [X]  I confirm that I read and will adhere to the JOSS code of conduct.
- [X] Repository: Is the source code for this software available at the repository url?
- [X] License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
**Contribution and authorship:**
- [X] Has the submitting author (@spinicist) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
- [X] Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines
**Example usage:**
- [ ] Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
 Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
**Automated tests:**
- [ ] Are there automated tests or manual steps described so that the functionality of the software can be verified?
**Community guidelines:**
- [X] Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
- [ ] Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
- [ ] A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
- [ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?
- [ ] Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
- [ ] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
