+++ docs/recon.rst
@@ -10,15 +10,15 @@ The ``recon`` command provides basic non-iterative reconstructions. This is usef
 
 By default ``recon`` will output complex-valued images in ``.h5`` format that have been combined using channel sensitivities extracted from the scan. If you are only interested in the magnitude images, add ``--mag`` to the command-line. If you only want a root-sum-squares reconstruction, add ``--rss``. If you would like NIFTI images as output, add ``--oft=nii``.
 
-Non-cartesian MRI data requires a Non-Uniform FFT (NUFFT) instead of a simple FFT for conversion between k-space and image space. The NUFFT consists of a gridding step and then a normal FFT. The gridding step in RIESLING is controlled by the oversampling and kernel options.
+Non-cartesian MRI data requires a Non-Uniform FFT (NUFFT) instead of a simple FFT for conversion between k-space and image space. The NUFFT consists of a gridding step and then a normal FFT [1, 2]. The gridding step in RIESLING is controlled by the oversampling and kernel options.
 
-``--os=X`` controls the grid oversampling ratio. The default value is 2, but this is very memory intensive (as it is a 3D grid, 2x oversampling requires 8 times the memory of the native matrix size). Reducing the over-sampling to a value of 1.3 leads to a reduced memory foot-print for little impact in image quality. 
+``--os=X`` controls the grid oversampling ratio. The default value is 2, but this is very memory intensive (as it is a 3D grid, 2x oversampling requires 8 times the memory of the native matrix size). Reducing the over-sampling to a value of 1.3 leads to a reduced memory foot-print for little impact in image quality [3]. 
 
-``--kb`` selects a Kaiser-Bessel kernel instead of the default nearest-neighbour kernel. Kaiser-Bessel is the default in most other toolboxes and may become the default in RIESLING in the future. The improvement in image quality from KB over NN is less marked in 3D imaging compared to 2D. The ``--kw`` option controls the kernel width, the default value is 3 and is sufficient for most applications.
+``--kb`` selects a Kaiser-Bessel (KB) kernel instead of the default nearest-neighbour (NN)kernel [3, 4]. Kaiser-Bessel is the default in most other toolboxes and may become the default in RIESLING in the future. The ``--kw`` option controls the kernel width, the default value is 3 and is sufficient for most applications.
 
-The final gridding option is ``--fast-grid``. RIESLING uses a parallelized gridding operation. In order to avoid multiple threads writing to the same Cartesian k-space point, the trajectory is sorted by Cartesian k-space location and threads each process their own chunk of the sorted co-ordinates. However, there may still be race conditions at the edge of each chunk, particularly for small images with highly oversampled k-space centers. To prevent this, each thread uses it's own workspace, and then these are combined into the final grid at the end. This process is thread-safe but doubles the memory requirements for the gridding operations. The ``--fast-grid`` option makes the threads write directly into the final grid, reducing memory consumption but at the risk of some writes into the grid being overwritten. When gridding high resolution images on a small number of threads, e.g. 10 or fewer, the probability of a race condition is vanishingly small. Use at your own risk.
[…]
