### 2021-09-14 c7edaa2
.rst interprets I. A. as lists not initials
