### 2021-09-14 spinicist/riesling#28 @spinicist
The standard gridding+FFT method is implemented in `riesling`. The nearest-neighbour gridding in Oesterle et al is equivalent to gridding with a top-hat kernel. The Kaiser-Bessel kernel is implemented with the formulas in Beatty et al.

I have removed the reference to Fessler's paper as this method is definitely not implemented, apologies for the confusion.  I have added the Jackson et al reference, and also the recent paper of Baron et al for Toeplitz embedding which in turn references the ISMRM abstract suggested. I have added citation markers in the text to make it clearer what is being referred to.
