### 2021-09-23 abca614
Bump Github CI to Ubuntu 18.04
