### 2021-09-23 spinicist/riesling#27 @spinicist
I asked for some help on the CMake Discourse and have made another attempt at fixing this on the `stdfs` branch. Sadly Github Actions does not have GCC 8 easily available to test against (only 7, 9 & 10, I have no idea why 8 is missing) so I can't properly test this - I have tested against GCC 9 and that still compiles fine.
