### 2021-09-23 e207aaf
Added Mapping calculation to Trajectory
