+++ src/main_sdc.cpp
@@ -20,7 +20,7 @@ int main_sdc(args::Subparser &parser)
     Kernel *kernel =
         kb ? (Kernel *)new KaiserBessel(kw.Get(), osamp.Get(), (info.type == Info::Type::ThreeD))
            : (Kernel *)new NearestNeighbour(kw ? kw.Get() : 1);
-    Gridder gridder(traj, osamp.Get(), kernel, fastgrid, log);
+    Gridder gridder(traj.mapping(osamp.Get(), kernel->radius()), kernel, fastgrid, log);
     dc = SDC::Pipe(traj, gridder, log);
   } else if (sdc.Get() == "radial") {
     dc = SDC::Radial(traj, log);
