+++ src/main_plan.cpp
@@ -21,7 +21,7 @@ int main_plan(args::Subparser &parser)
   Kernel *kernel =
       kb ? (Kernel *)new KaiserBessel(3, osamp.Get(), (traj.info().type == Info::Type::ThreeD))
          : (Kernel *)new NearestNeighbour();
-  Gridder gridder(traj, osamp.Get(), kernel, fastgrid, log);
+  Gridder gridder(traj.mapping(osamp.Get(), kernel->radius()), kernel, fastgrid, log);
   Cx4 grid4 = gridder.newMultichannel(traj.info().channels);
   Cx4 grid3 = gridder.newMultichannel(1);
   FFT::ThreeDMulti fft3(grid3, log);
