+++ src/main_traj.cpp
@@ -21,7 +21,7 @@ int main_traj(args::Subparser &parser)
   Kernel *kernel =
       kb ? (Kernel *)new KaiserBessel(kw.Get(), osamp.Get(), (info.type == Info::Type::ThreeD))
          : (Kernel *)new NearestNeighbour(kw ? kw.Get() : 1);
-  Gridder gridder(traj, osamp.Get(), kernel, fastgrid, log);
+  Gridder gridder(traj.mapping(osamp.Get(), kernel->radius()), kernel, fastgrid, log);
   SDC::Load(sdc.Get(), traj, gridder, log);
   Cx4 grid = gridder.newMultichannel(1);
   FFT::ThreeDMulti fft(grid, log);
