+++ test/gridder.cpp
@@ -1,19 +1,21 @@
-#include "../src/gridder.h"
 #include "../src/log.h"
+#include "../src/op/grid-kb.h"
+#include "../src/op/grid-nn.h"
 #include "../src/trajectory.h"
 #include <catch2/catch.hpp>
 
-TEST_CASE("Gridder with single point", "GRID-SINGLE")
+TEST_CASE("Gridding", "GRIDDING")
 {
   Log log;
-  Info info{.type = Info::Type::ThreeD,
-            .channels = 4,
-            .matrix = {2, 2, 2},
-            .read_points = 1,
-            .read_gap = 0,
-            .spokes_hi = 1,
-            .spokes_lo = 0,
-            .lo_scale = 1};
+  Info info{
+      .type = Info::Type::ThreeD,
+      .channels = 4,
+      .matrix = {2, 2, 2},
+      .read_points = 1,
+      .read_gap = 0,
+      .spokes_hi = 1,
+      .spokes_lo = 0,
+      .lo_scale = 1};
   float const osamp = 2.f;
   R3 points(3, 1, 1);
   points.setZero();
@@ -21,8 +23,7 @@ TEST_CASE("Gridder with single point", "GRID-SINGLE")
 
   SECTION("NN")
   {
-    NearestNeighbour kernel;
-    Gridder gridder(traj.mapping(osamp, kernel.radius()), &kernel, false, log);
+    GridNN gridder(traj, osamp, false, log);
     Cx3 rad = info.noncartesianVolume();
     CHECK(rad.dimension(0) == 4);
     CHECK(rad.dimension(1) == 1);
@@ -32,16 +33,15 @@ TEST_CASE("Gridder with single point", "GRID-SINGLE")
     CHECK(cart.dimension(3) == 4);
     rad.setConstant(1.f);
     cart.setZero();
-    gridder.toCartesian(rad, cart);
+    gridder.Adj(rad, cart);
     CHECK(cart(0, 2, 2, 2).real() == Approx(1.f));
-    gridder.toNoncartesian(cart, rad);
+    gridder.A(cart, rad);
     CHECK(rad(0, 0, 0).real() == Approx(1.f));
   }
 
   SECTION("NN Multicoil")
   {
-    NearestNeighbour kernel;
-    Gridder gridder(traj.mapping(osamp, kernel.radius()), &kernel, false, log);
+    GridNN gridder(traj, osamp, false, log);
     Cx3 rad = info.noncartesianVolume();
     CHECK(rad.dimension(0) == info.channels);
     CHECK(rad.dimension(1) == info.read_points);
@@ -52,12 +52,12 @@ TEST_CASE("Gridder with single point", "GRID-SINGLE")
     CHECK(cart.dimension(3) == 4);
     rad.setConstant(1.f);
     cart.setZero();
-    gridder.toCartesian(rad, cart);
+    gridder.Adj(rad, cart);
     CHECK(cart(0, 2, 2, 2).real() == Approx(1.f));
     CHECK(cart(1, 2, 2, 2).real() == Approx(1.f));
     CHECK(cart(2, 2, 2, 2).real() == Approx(1.f));
     CHECK(cart(3, 2, 2, 2).real() == Approx(1.f));
-    gridder.toNoncartesian(cart, rad);
+    gridder.A(cart, rad);
     CHECK(rad(0, 0, 0).real() == Approx(1.f));
     CHECK(rad(1, 0, 0).real() == Approx(1.f));
     CHECK(rad(2, 0, 0).real() == Approx(1.f));
@@ -66,9 +66,7 @@ TEST_CASE("Gridder with single point", "GRID-SINGLE")
 
[…]
