+++ src/kernels.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#include "kernel_kb.h"
-#include "kernel_nn.h"
