+++ src/espirit.cpp
@@ -3,16 +3,16 @@
 #include "cropper.h"
 #include "decomp.h"
 #include "fft_plan.hpp"
-#include "gridder.h"
 #include "hankel.h"
 #include "io_nifti.h"
+#include "op/grid.h"
 #include "padder.h"
 #include "tensorOps.h"
 #include "threads.h"
 #include "vc.h"
 
 Cx4 ESPIRIT(
-    Gridder const &gridder,
+    std::unique_ptr<GridOp> const &gridder,
     Cx3 const &data,
     long const kRad,
     long const calRad,
@@ -23,9 +23,9 @@ Cx4 ESPIRIT(
   log.info(FMT_STRING("ESPIRIT Calibration Radius {} Kernel Radius {}"), calRad, kRad);
 
   log.info(FMT_STRING("Calculating k-space kernels"));
-  Cx4 grid = gridder.newMultichannel(data.dimension(0)); // Maps will end up here
-  R3 valsImage(gridder.gridDims());
-  gridder.toCartesian(data, grid);
+  Cx4 grid = gridder->newMultichannel(data.dimension(0)); // Maps will end up here
+  R3 valsImage(gridder->gridDims());
+  gridder->Adj(data, grid);
   Cx5 const all_kernels = ToKernels(grid, kRad, calRad, gap, log);
   Cx5 const mini_kernels = LowRankKernels(all_kernels, thresh, log);
   long const retain = mini_kernels.dimension(4);

[…]
