+++ test/op/crop.cpp
@@ -9,7 +9,7 @@ TEST_CASE("ops-crop", "[ops]")
 
   SECTION("Dot Test")
   {
-    long const cropSz = 8;
+    long const cropSz = 7;
     Cx3 x(fullSz, fullSz, fullSz), xy(cropSz, cropSz, cropSz), y(cropSz, cropSz, cropSz),
         yx(fullSz, fullSz, fullSz);
 
@@ -20,6 +20,9 @@ TEST_CASE("ops-crop", "[ops]")
     crop.A(x, xy);
     crop.Adj(y, yx);
 
-    CHECK(Dot(x, yx) == Dot(xy, y));
+    // Don't forget conjugate on second dot argument
+    auto const xx = Dot(x, yx);
+    auto const yy = Dot(xy, y);
+    CHECK(std::abs((yy - xx) / (yy + xx + 1.e-15f)) == Approx(0).margin(1.e-6));
   }
 }
