### 2021-10-06 bfb8803
Update CI Publish Ubuntu version
