+++ .github/workflows/build.yml
@@ -125,7 +125,7 @@ jobs:
         config:
         - {
           name: "Ubuntu GCC", artifact: "riesling-linux.tar.gz",
-          os: ubuntu-16.04
+          os: ubuntu-18.04
         }
         - {
           name: "macOS", artifact: "riesling-macos.tar.gz",
