### 2021-10-06 #3500 @spinicist
I have addressed the problem in the example repository by updating the example into a Notebook, which now runs successfully in MyBinder (https://hub.gke2.mybinder.org/user/spinicist-riesling-examples-xaoj7os3/lab/tree/rrsg-cgsense.ipynb).
