+++ docs/paper.bib
@@ -4,6 +4,7 @@
     year = {2021},
     pages = {21},
     abstract = {Magnetic Resonance Imaging (MRI) scanners produce loud acoustic noise originating from vibrational Lorentz forces induced by rapidly changing currents in the magnetic field gradient coils. Using zero echo time (ZTE) MRI pulse sequences, gradient switching can be reduced to a minimum, which enables near silent operation. Besides silent MRI, ZTE offers further interesting characteristics, including a nominal echo time of TE = 0 (thus capturing short-lived signals from MR tissues which are otherwise MRinvisible), 3D radial sampling (providing motion robustness), and ultra-short repetition times (providing fast and efficient scanning). In this work we describe the main concepts behind ZTE imaging with a focus on conceptual understanding of the imaging sequences, relevant acquisition parameters, commonly observed image artefacts, and image contrasts. We will further describe a range of methods for anatomical and functional neuroimaging, together with recommendations for successful implementation.},
+    doi = {10.1016/j.pnmrs.2021.03.002},
     journal = {Progress in Nuclear Magnetic Resonance Spectroscopy},
     language = {en}
 }
