### 2021-10-06 #3500 @emdupre
Thank you, @spinicist !

To quickly confirm : the Zenodo metadata lists the authors as Tobias Wood; Emil Ljungberg; Martin Krämer while the repository README (and paper) lists the authors as Tobias C Wood, Emil Ljungberg, Florian Wiesinger.

Can we ensure that Florian Wiesinger is included as an author in the Zenodo archive, given his authorship on the paper ? Separately, can you confirm that you intended to list Martin Krämer as an author on the Zenodo archive ?
