### 2021-10-09 3e0edbc
Implement ReconOp
