+++ src/op/grid-basis-kb.cpp
@@ -1,4 +1,4 @@
-#include "grid-kb.h"
+#include "grid-basis-kb.h"
 
 #include "../cropper.h"
 #include "../fft_plan.h"
@@ -10,14 +10,15 @@
 #include <cmath>
 
 template <int InPlane, int ThroughPlane>
-GridKB<InPlane, ThroughPlane>::GridKB(
+GridBasisKB<InPlane, ThroughPlane>::GridBasisKB(
     Trajectory const &traj,
     float const os,
     bool const unsafe,
+    R2 &basis,
     Log &log,
     float const inRes,
     bool const shrink)
-    : GridOp(traj.mapping(os, (InPlane / 2), inRes, shrink), unsafe, log)
+    : GridBasisOp(traj.mapping(os, (InPlane / 2), inRes, shrink), unsafe, basis, log)
     , betaIn_{(float)M_PI *
               sqrtf(pow(InPlane * (mapping_.osamp - 0.5f) / mapping_.osamp, 2.f) - 0.8f)}
     , betaThrough_{(float)M_PI *
@@ -29,25 +30,6 @@ GridKB<InPlane, ThroughPlane>::GridKB(
   std::iota(indIn_.data(), indIn_.data() + InPlane, -InPlane / 2);
   std::iota(indThrough_.data(), indThrough_.data() + ThroughPlane, -ThroughPlane / 2);
 }
-
-template <int InPlane, int ThroughPlane>
-GridKB<InPlane, ThroughPlane>::GridKB(
-    Mapping const &mapping,
-    bool const unsafe,
-    Log &log)
-    : GridOp(mapping, unsafe, log)
-    , betaIn_{(float)M_PI *
-              sqrtf(pow(InPlane * (mapping_.osamp - 0.5f) / mapping_.osamp, 2.f) - 0.8f)}
-    , betaThrough_{(float)M_PI *
-                   sqrtf(pow(ThroughPlane * (mapping_.osamp - 0.5f) / mapping_.osamp, 2.f) - 0.8f)}
-    , fft_{Sz3{InPlane, InPlane, ThroughPlane}, Log(), 1}
-{
-
-  // Array of indices used when building the kernel
-  std::iota(indIn_.data(), indIn_.data() + InPlane, -InPlane / 2);
-  std::iota(indThrough_.data(), indThrough_.data() + ThroughPlane, -ThroughPlane / 2);
-}
-
 
 template <int W, typename T>
 inline decltype(auto) KB(T const &x, float const beta)
@@ -60,7 +42,7 @@ inline decltype(auto) KB(T const &x, float const beta)
 }
 
 template <int InPlane, int ThroughPlane>
-void GridKB<InPlane, ThroughPlane>::kernel(Point3 const r, float const dc, Kernel &k) const
+void GridBasisKB<InPlane, ThroughPlane>::kernel(Point3 const r, float const dc, Kernel &k) const
 {
   InPlaneArray const kx = KB<InPlane>(indIn_.constant(r[0]) - indIn_, betaIn_);
   InPlaneArray const ky = KB<InPlane>(indIn_.constant(r[1]) - indIn_, betaIn_);
@@ -86,44 +68,56 @@ void GridKB<InPlane, ThroughPlane>::kernel(Point3 const r, float const dc, Kerne
 }
 
 template <int InPlane, int ThroughPlane>
-void GridKB<InPlane, ThroughPlane>::Adj(Cx3 const &noncart, Cx4 &cart) const
+void GridBasisKB<InPlane, ThroughPlane>::Adj(Output const &noncart, Input &cart) const
 {
-  assert(noncart.dimension(0) == cart.dimension(0));
[…]
