### 2021-10-12 17aa900
Fix echoes bug, update Matlab demo
