+++ matlab/riesling_read.m
@@ -12,26 +12,26 @@ function varargout = riesling_read(fname)
 % Emil Ljungberg, King's College London, 2021
 % Martin Krämer, University Hospital Jena, 2021
 
-% get h5 file info and check if image data is present
+% always read info 
+info = h5read(fname, '/info');
+
+% Open h5 file
 file_info = h5info(fname);
 is_img_data = any(contains({file_info.Datasets.Name}, 'image'));
-
 if is_img_data % load only image data
     data = h5read(fname, '/image');
     img = data.r + 1j*data.i;
     
     varargout{1} = img;
+    varargout{2} = info;
 else % load k-space data and trajectory
     traj = h5read(fname, '/trajectory');
     data = h5read(fname, '/noncartesian');
     kspace = data.r + 1j*data.i;
 
     varargout{1} = kspace;
-    varargout{3} = traj;
+    varargout{2} = traj;
+    varargout{3} = info;
 end
-
-% always read info 
-info = h5read(fname, '/info');
-varargout{2} = info;
 
 end
