### 2021-10-15 10089c1
Info struct checks and demo update
