### 2021-10-15 b9bfb81
Remove unused variable
