### 2021-10-19 e59a355
Fix maximum gridding radius
