+++ src/sim.h
@@ -8,3 +8,10 @@ struct SimResult
   Eigen::MatrixXf dynamics;
   Eigen::MatrixXf parameters;
 };
+
+// Another simple struct for passing around the main ZTE sequence parameters
+struct Sequence
+{
+  long sps;
+  float alpha, TR, TI, Trec;
+};
