+++ src/sim-eddy.h
@@ -11,14 +11,10 @@ SimResult Diffusion(
     float const T1Lo,
     float const T1Hi,
     long const nbeta,
-    long const ngamma,
     float const betaLo,
     float const betaHi,
-    long const sps,
-    float const alpha,
-    float const TR,
-    float const TI,
-    float const Trec,
+    long const ngamma,
+    Sequence const seq,
     Log &log);
 
 }
