+++ src/sim-prep.cpp
@@ -11,57 +11,140 @@ SimResult Simple(
     long const nbeta,
     float const betaLo,
     float const betaHi,
-    long const sps,
-    float const alpha,
-    float const TR,
-    float const TI,
-    float const Trec,
+    bool const betaLog,
+    Sequence const seq,
     Log &log)
 {
   log.info("Basic MP-ZTE simulation");
-  log.info(FMT_STRING("Seg length {}, FA {}, TR {}s, TI {}s, Trec {}s"), sps, alpha, TR, TI, Trec);
+  log.info(
+      FMT_STRING("SPS {}, FA {}, TR {}s, TI {}s, Trec {}s"),
+      seq.sps,
+      seq.alpha,
+      seq.TR,
+      seq.TI,
+      seq.Trec);
   log.info(FMT_STRING("{} values of T1 from {} to {}s"), nT1, T1Lo, T1Hi);
   log.info(FMT_STRING("{} values of β from {} to {}"), nbeta, betaLo, betaHi);
   long totalN = nT1 * nbeta;
-  Eigen::MatrixXf sims(totalN, sps); // SVD expects observations in rows
+  Eigen::MatrixXf sims(totalN, seq.sps); // SVD expects observations in rows
   Eigen::MatrixXf parameters(totalN, 2);
 
-  float const cosa = cos(alpha * M_PI / 180.f);
+  float const cosa = cos(seq.alpha * M_PI / 180.f);
   Eigen::Matrix2f A;
   A << cosa, 0.f, 0.f, 1.f;
-  for (long ib = 0; ib < nbeta; ib++) {
-    float const fib = ib / (nbeta - 1.f);
-    float const beta = betaLo * (1.f - fib) + betaHi * fib;
-    Eigen::Matrix2f B;
-    B << beta, 0.f, 0.f, 1.f;
-    for (long it = 0; it < nT1; it++) {
-      // Set up matrices
-      float const fit = it / (nT1 - 1.f);
-      float const R1 = 1.f / (T1Lo * (1.f - fit) + T1Hi * fit);
-      float const e1 = exp(-R1 * TR);
-      Eigen::Matrix2f E1;
-      E1 << e1, 1 - e1, 0.f, 1.f;
-      float const einv = exp(-R1 * TI);
-      Eigen::Matrix2f Ei;
-      Ei << einv, 1 - einv, 0.f, 1.f;
-      float const erec = exp(-R1 * Trec);
-      Eigen::Matrix2f Er;
-      Er << erec, 1 - erec, 0.f, 1.f;
+  for (long it = 0; it < nT1; it++) {
+    // Set up matrices
+    float const fit = it / (nT1 - 1.f);
+    float const T1 = T1Lo * std::pow(T1Hi / T1Lo, fit);
+    float const R1 = 1.f / T1;
+    float const e1 = exp(-R1 * seq.TR);
+    Eigen::Matrix2f E1;
+    E1 << e1, 1 - e1, 0.f, 1.f;
+    float const einv = exp(-R1 * seq.TI);
+    Eigen::Matrix2f Ei;
+    Ei << einv, 1 - einv, 0.f, 1.f;
+    float const erec = exp(-R1 * seq.Trec);
+    Eigen::Matrix2f Er;
+    Er << erec, 1 - erec, 0.f, 1.f;
+
+    for (long ib = 0; ib < nbeta; ib++) {
+      float const fib = ib / (nbeta - 1.f);
+      float const ref = 0.1;
+      float const beta = betaLog ? (ref * std::pow((betaHi - betaLo) / ref, fib) - (ref - betaLo))
+                                 : (betaLo + fib * (betaHi - betaLo));
[…]
