### 2021-10-20 f8e04b3
Improve output of basis/parameter images
