### 2021-10-21 6554ec3
Add B1 to sims and fix scaling
