+++ src/sim-eddy.cpp
@@ -5,42 +5,37 @@
 
 namespace Sim {
 
-SimResult Diffusion(
-    long const nT1,
-    float const T1Lo,
-    float const T1Hi,
-    long const nbeta,
-    float const betaLo,
-    float const betaHi,
-    long const ngamma,
-    Sequence const seq,
-    Log &log)
+Result Eddy(
+  Parameter const T1p,
+  Parameter const betap,
+  Parameter const gammap,
+  Parameter const B1p,
+  bool const negPC,
+  Sequence const seq,
+  Log &log)
 {
   log.info("Eddy Current MP-ZTE simulation");
   log.info(
-      FMT_STRING("SPS {}, FA {}, TR {}s, TI {}s, Trec {}s"),
-      seq.sps,
-      seq.alpha,
-      seq.TR,
-      seq.TI,
-      seq.Trec);
-  log.info(FMT_STRING("{} values of T1 from {} to {}s"), nT1, T1Lo, T1Hi);
-  log.info(FMT_STRING("{} values of β from {} to {}"), nbeta, betaLo, betaHi);
-  log.info(FMT_STRING("{} values of ɣ from 0 to 2π"), ngamma);
-  long totalN = nT1 * nbeta * ngamma;
+    FMT_STRING("SPS {}, FA {}, TR {}s, TI {}s, Trec {}s"),
+    seq.sps,
+    seq.alpha,
+    seq.TR,
+    seq.TI,
+    seq.Trec);
+  log.info(FMT_STRING("{} values of T1 from {} to {}s"), T1p.N, T1p.lo, T1p.hi);
+  log.info(FMT_STRING("{} values of β from {} to {}"), betap.N, betap.lo, betap.hi);
+  log.info(FMT_STRING("{} values of ɣ from {} to {}"), gammap.N, gammap.lo, gammap.hi);
+  log.info(FMT_STRING("{} values of B1 from {} to {}"), B1p.N, B1p.lo, B1p.hi);
+  long totalN = T1p.N * betap.N * gammap.N * B1p.N;
   Eigen::MatrixXf sims(totalN, 4 * seq.sps); // SVD expects observations in rows
-  Eigen::MatrixXf parameters(totalN, 3);
+  Eigen::MatrixXf parameters(totalN, 4);
 
-  float const cosa = cos(seq.alpha * M_PI / 180.f);
-  Eigen::Matrix2f A;
-  A << cosa, 0.f, 0.f, 1.f;
   auto task = [&](long const lo, long const hi, long const ti) {
-    long row = lo * ngamma * nbeta;
+    long row = lo * gammap.N * betap.N * B1p.N;
     for (long it = lo; it < hi; it++) {
       log.progress(it, lo, hi);
       // Set up matrices
-      float const fit = it / (nT1 - 1.f);
-      float const T1 = T1Lo * std::pow(T1Hi / T1Lo, fit);
+      float const T1 = T1p.value(it);
       float const R1 = 1.f / T1;
       float const e1 = exp(-R1 * seq.TR);
       Eigen::Matrix2f E1;
@@ -51,64 +46,73 @@ SimResult Diffusion(
       float const erec = exp(-R1 * seq.Trec);
       Eigen::Matrix2f Er;
       Er << erec, 1 - erec, 0.f, 1.f;
-      for (long ib = 0; ib < nbeta; ib++) {
-        float const fib = ib / (nbeta - 1.f);
-        float const beta = betaLo * std::pow(betaHi / betaLo, fib);
-        Eigen::Matrix2f B;
-        B << beta, 0.f, 0.f, 1.f;
-        for (long ig = 0; ig < ngamma; ig++) {
[…]
