+++ src/sim-eddy.h
@@ -6,15 +6,13 @@
 
 namespace Sim {
 
-SimResult Diffusion(
-    long const nT1,
-    float const T1Lo,
-    float const T1Hi,
-    long const nbeta,
-    float const betaLo,
-    float const betaHi,
-    long const ngamma,
-    Sequence const seq,
-    Log &log);
+Result Eddy(
+  Parameter const T1,
+  Parameter const beta,
+  Parameter const gamma,
+  Parameter const B1,
+  bool const negPC,
+  Sequence const seq,
+  Log &log);
 
 }
