+++ src/main_basis_sim.cpp
@@ -18,60 +18,64 @@ int main_basis_sim(args::Subparser &parser)
       parser, "TI", "Inversion time (from prep to segment start)", {"ti"}, 0.45f);
   args::ValueFlag<float> Trec(
       parser, "TREC", "Recover time (from segment end to prep)", {"trec"}, 0.f);
+  args::ValueFlag<long> nT1(
+      parser, "N", "Number of T1 values for basis (default 128)", {"nT1"}, 32);
   args::ValueFlag<float> T1Lo(
       parser, "T1", "Low value of T1 for basis (default 0.4s)", {"T1lo"}, 0.4f);
   args::ValueFlag<float> T1Hi(
       parser, "T1", "High value of T1 for basis (default 5s)", {"T1hi"}, 5.f);
-  args::ValueFlag<long> nT1(
-      parser, "N", "Number of T1 values for basis (default 128)", {"nT1"}, 32);
-  args::ValueFlag<float> bLo(parser, "β", "Low value for β (default -1)", {"betalo"}, -1.f);
-  args::ValueFlag<float> bHi(parser, "β", "High value for β (default 1)", {"betahi"}, 1.f);
   args::ValueFlag<long> nb(
       parser, "N", "Number of beta values for basis (default 128)", {"nbeta"}, 32);
+  args::ValueFlag<float> bLo(parser, "β", "Low value for β (default -1)", {"betalo"}, -1.f);
+  args::ValueFlag<float> bHi(parser, "β", "High value for β (default 1)", {"betahi"}, 1.f);
   args::Flag bLog(parser, "", "Use logarithmic spacing for β", {"betalog"});
+
+  args::ValueFlag<long> nB1(parser, "N", "Number of B1 values for basis (default 7)", {"nB1"}, 7);
+  args::ValueFlag<float> B1Lo(parser, "B1", "Low value for B1 (default 0.7)", {"B1lo"}, 0.7f);
+  args::ValueFlag<float> B1Hi(parser, "B1", "High value for B1 (default 1.3)", {"B1hi"}, 1.3f);
+
   args::ValueFlag<long> ng(parser, "N", "Number of eddy-current angles", {"eddy"}, 32);
-  args::Flag pc(parser, "PC", "Uses 0/180 phase-cycling", {"pc"});
+  args::ValueFlag<float> gLo(parser, "ɣ", "Low value for eddy-current angles (default -π)", {"eddylo"}, -M_PI);
+  args::ValueFlag<float> gHi(parser, "ɣ", "High value for eddy-current angles (default π)", {"eddyhi"}, M_PI);
+  args::Flag negPC(parser, "-PC", "Phase-cycling increment is negative", {"negpc"});
   args::ValueFlag<float> thresh(
-      parser, "T", "Threshold for SVD retention (default 0.05)", {"thresh"}, 0.05);
+      parser, "T", "Threshold for SVD retention (default 95%)", {"thresh"}, 95.f);
   args::ValueFlag<long> nBasis(
       parser, "N", "Number of basis vectors to retain (overrides threshold)", {"nbasis"}, 0);
 
   Log log = ParseCommand(parser);
 
-  Sequence const seq{sps.Get(), alpha.Get(), TR.Get(), TI.Get(), Trec.Get()};
-  SimResult results;
+  Sim::Sequence const seq{sps.Get(), alpha.Get(), TR.Get(), TI.Get(), Trec.Get()};
+  Sim::Parameter const T1{nT1.Get(), T1Lo.Get(), T1Hi.Get(), true};
+  Sim::Parameter const beta{nb.Get(), bLo.Get(), bHi.Get(), bLog};
[…]
