+++ src/op/grid.h
@@ -1,10 +1,11 @@
 #pragma once
 
 #include "../trajectory.h"
+#include "gridBase.h"
 #include "operator.h"
 #include <memory>
 
-struct GridOp : Operator<4, 3>
+struct GridOp : Operator<4, 3>, GridBase
 {
   GridOp(Mapping map, bool const unsafe, Log &log);
   virtual ~GridOp(){};
@@ -15,23 +16,10 @@ struct GridOp : Operator<4, 3>
   Input::Dimensions inSize() const;
   Output::Dimensions outputDimensions() const;
 
-  Sz3 gridDims() const;                        // Returns the dimensions of the grid
   Cx4 newMultichannel(long const nChan) const; // Returns a correctly sized multi-channel grid
-  void setSDCExponent(float const dce);
-  void setSDC(float const dc);
-  void setSDC(R2 const &sdc);
-  void setUnsafe();
-  void setSafe();
-  void sqrtOn(); // Use square-root of gridding kernel for Pipe SDC
-  void sqrtOff();
   virtual R3 apodization(Sz3 const sz) const = 0; // Calculate the apodization factor for this grid
-  Mapping const &mapping() const;
 
 protected:
-  Mapping mapping_;
-  bool safe_, sqrt_;
-  Log &log_;
-  float DCexp_;
 };
 
 std::unique_ptr<GridOp> make_grid(
@@ -42,6 +30,7 @@ std::unique_ptr<GridOp> make_grid(
     Log &log,
     float const res = -1.f,
     bool const shrink = false);
+
 std::unique_ptr<GridOp> make_grid(
     Mapping const &mapping,
     bool const kb,
