+++ src/op/grid-basis-nn.h
@@ -13,6 +13,11 @@ struct GridBasisNN final : GridBasisOp
       Log &log,
       float const inRes = -1.f,
       bool const shrink = false);
+  GridBasisNN(
+      Mapping const &map,
+      bool const unsafe,
+      R2 &basis,
+      Log &log);
 
   void A(Input const &x, Output &y) const;
   void Adj(Output const &x, Input &y) const;
