+++ src/main_defs.h
@@ -4,6 +4,7 @@
 
 int main_basis_admm(args::Subparser &parser);
 int main_basis_blend(args::Subparser &parser);
+int main_basis_cg(args::Subparser &parser);
 int main_basis_dict(args::Subparser &parser);
 int main_basis_recon(args::Subparser &parser);
 int main_basis_sim(args::Subparser &parser);
