+++ src/sim-mupa.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "log.h"
+#include "sim.h"
+#include "types.h"
+
+namespace Sim {
+
+Result MUPA(
+  Parameter const T1,
+  Parameter const T2,
+  Parameter const B1,
+  Sequence const seq,
+  Log &log);
+
+}
