### 2021-10-26 af96cd1
Add a benchmark for dot product
