### 2021-10-26 2c3421b
Parallelize tensorOps
