+++ src/sim-mupa.cpp
@@ -25,7 +25,7 @@ Result MUPA(
   log.info(FMT_STRING("{} values of B1 from {} to {}"), B1p.N, B1p.lo, B1p.hi);
   long totalN = T1p.N * T2p.N * B1p.N;
   Eigen::MatrixXf sims(totalN, 4 * seq.sps); // SVD expects observations in rows
-  Eigen::MatrixXf parameters(totalN, 3);
+  Eigen::MatrixXf parameters(totalN, 4);
 
   Eigen::Matrix2f inv;
   inv << -1.f, 0.f,
@@ -37,17 +37,18 @@ Result MUPA(
       // Set up matrices
       float const T1 = T1p.value(it);
       float const R1 = 1.f / T1;
+      Eigen::Matrix2f E1, Eramp, Essi, Ei, Er;
       float const e1 = exp(-R1 * seq.TR);
-      Eigen::Matrix2f E1;
+      float const eramp = exp(-R1 * seq.Tramp);
+      float const essi = exp(-R1 * seq.Tssi);
+      float const einv = exp(-R1 * seq.TI);
       E1 << e1, 1 - e1,
             0.f, 1.f;
-      float const einv = exp(-R1 * seq.TI);
-      Eigen::Matrix2f Ei;
-      Ei << einv, 1 - einv,
+      Eramp << eramp, 1 - eramp,
             0.f, 1.f;
-      float const erec = exp(-R1 * seq.Trec);
-      Eigen::Matrix2f Er;
-      Er << erec, 1 - erec,
+      Essi << essi, 1 - essi,
+            0.f, 1.f;
+      Ei << einv, 1 - einv,
             0.f, 1.f;
 
       for (long iB1 = 0; iB1 < B1p.N; iB1++) {
@@ -67,9 +68,9 @@ Result MUPA(
                 0.f, 1.f;
 
           // Get steady state after prep-pulse for first segment
-          Eigen::Matrix2f const seg = (E1 * A).pow(seq.sps);
+          Eigen::Matrix2f const seg = Essi * Eramp * (E1 * A).pow(seq.sps) * Eramp;
           Eigen::Matrix2f const SS =
-              E2 * seg * seg * seg * Ei * inv * seg;
+              Eramp * Essi * E2 * seg * seg * seg * Ei * inv * Essi * seg;
           float const Mz_ss = SS(0, 1) / (1.f - SS(0, 0));
 
           // Now fill in dynamic
@@ -80,20 +81,22 @@ Result MUPA(
               Mz = A * Mz;
               Mz = E1 * Mz;
           }
-          Mz = Ei * inv * Mz;
+          Mz = Eramp * Ei * inv * Essi * Eramp * Mz;
           for (long is = 0; is < 3; is++) {
             for (long ii = 0; ii < seq.sps; ii++) {
                 sims(row, col++) = Mz(0) * sina;
                 Mz = A * Mz;
                 Mz = E1 * Mz;
             }
+            Mz = Eramp * Essi * Eramp * Mz;
           }
           if (col != (4 * seq.sps)) {
               Log::Fail("Programmer error");
           }
-          parameters(row, 0) = T1;
-          parameters(row, 1) = T2;
-          parameters(row, 2) = B1;
+          parameters(row, 0) = Mz_ss;
+          parameters(row, 1) = T1;
+          parameters(row, 2) = T2;
+          parameters(row, 3) = B1;
           row++;
         }
       }
