+++ src/op/gridBase.h
@@ -9,7 +9,7 @@ struct GridBase
 
   virtual R3 apodization(Sz3 const sz) const = 0; // Calculate the apodization factor for this grid
 
-  Sz3 gridDims() const;                        // Returns the dimensions of the grid
+  Sz3 gridDims() const; // Returns the dimensions of the grid
   void setSDCExponent(float const dce);
   void setSDC(float const dc);
   void setSDC(R2 const &sdc);
