### 2021-11-02 91bde32
Stopped over-optimising basis
