+++ src/op/grid-basis.h
@@ -29,7 +29,7 @@ std::unique_ptr<GridBasisOp> make_grid_basis(
     float const os,
     bool const kb,
     bool const fastgrid,
-    R2 &basis,
+    R2 const &basis,
     Log &log,
     float const res = -1.f,
     bool const shrink = false);
@@ -38,5 +38,5 @@ std::unique_ptr<GridBasisOp> make_grid_basis(
     Mapping const &mapping,
     bool const kb,
     bool const fastgrid,
-    R2 &basis,
+    R2 const &basis,
     Log &log);
