+++ src/op/grid-basis-nn.h
@@ -9,14 +9,14 @@ struct GridBasisNN final : GridBasisOp
       Trajectory const &traj,
       float const os,
       bool const unsafe,
-      R2 &basis,
+      R2 const &basis,
       Log &log,
       float const inRes = -1.f,
       bool const shrink = false);
   GridBasisNN(
       Mapping const &map,
       bool const unsafe,
-      R2 &basis,
+      R2 const &basis,
       Log &log);
 
   void A(Input const &x, Output &y) const;
