+++ src/op/grid-basis-kb.h
@@ -11,14 +11,14 @@ struct GridBasisKB final : GridBasisOp
       Trajectory const &traj,
       float const os,
       bool const unsafe,
-      R2 &basis,
+      R2 const &basis,
       Log &log,
       float const inRes = -1.f,
       bool const shrink = false);
   GridBasisKB(
       Mapping const &mapping,
       bool const unsafe,
-      R2 &basis,
+      R2 const &basis,
       Log &log);
 
   void A(Input const &x, Output &y) const;
