### 2021-11-02 724992b
Write out phantom images
