+++ src/sim-mupa.cpp
@@ -6,106 +6,95 @@
 namespace Sim {
 
 Result MUPA(
-  Parameter const T1p,
-  Parameter const T2p,
-  Parameter const B1p,
-  Sequence const seq,
-  Log &log)
+    Parameter const T1p,
+    Parameter const T2p,
+    Parameter const B1p,
+    Sequence const seq,
+    long const nRand,
+    Log &log)
 {
   log.info("MUPA MP-ZTE simulation");
   log.info(
-    FMT_STRING("SPS {}, FA {}, TR {}s, TI {}s, Trec {}s"),
-    seq.sps,
-    seq.alpha,
-    seq.TR,
-    seq.TI,
-    seq.Trec);
+      FMT_STRING("SPS {}, FA {}, TR {}s, TI {}s, Trec {}s"),
+      seq.sps,
+      seq.alpha,
+      seq.TR,
+      seq.TI,
+      seq.Trec);
   log.info(FMT_STRING("{} values of T1 from {} to {}s"), T1p.N, T1p.lo, T1p.hi);
   log.info(FMT_STRING("{} values of T2 from {} to {}"), T2p.N, T2p.lo, T2p.hi);
   log.info(FMT_STRING("{} values of B1 from {} to {}"), B1p.N, B1p.lo, B1p.hi);
-  long totalN = T1p.N * T2p.N * B1p.N;
-  Eigen::MatrixXf sims(totalN, 4 * seq.sps); // SVD expects observations in rows
-  Eigen::MatrixXf parameters(totalN, 4);
+  ParameterGenerator<3> gen({T1p, T2p, B1p});
+  long totalN = (nRand > 0) ? nRand : gen.totalN();
+  Result result;
+  result.dynamics.resize(totalN, 4 * seq.sps);
+  result.parameters.resize(totalN, 3);
+  result.Mz_ss.resize(totalN);
 
   Eigen::Matrix2f inv;
-  inv << -1.f, 0.f,
-          0.f, 1.f;
+  inv << -1.f, 0.f, 0.f, 1.f;
   auto task = [&](long const lo, long const hi, long const ti) {
-    long row = lo * T2p.N * B1p.N;
-    for (long it = lo; it < hi; it++) {
-      log.progress(it, lo, hi);
+    for (long ip = lo; ip < hi; ip++) {
+      log.progress(ip, lo, hi);
       // Set up matrices
-      float const T1 = T1p.value(it);
+      auto const P = (nRand > 0) ? gen.rand() : gen.values(ip);
+      float const T1 = P(0);
       float const R1 = 1.f / T1;
       Eigen::Matrix2f E1, Eramp, Essi, Ei, Er;
       float const e1 = exp(-R1 * seq.TR);
       float const eramp = exp(-R1 * seq.Tramp);
       float const essi = exp(-R1 * seq.Tssi);
       float const einv = exp(-R1 * seq.TI);
-      E1 << e1, 1 - e1,
-            0.f, 1.f;
-      Eramp << eramp, 1 - eramp,
-            0.f, 1.f;
-      Essi << essi, 1 - essi,
-            0.f, 1.f;
-      Ei << einv, 1 - einv,
-            0.f, 1.f;
+      E1 << e1, 1 - e1, 0.f, 1.f;
+      Eramp << eramp, 1 - eramp, 0.f, 1.f;
+      Essi << essi, 1 - essi, 0.f, 1.f;
+      Ei << einv, 1 - einv, 0.f, 1.f;
 
-      for (long iB1 = 0; iB1 < B1p.N; iB1++) {
-        float const B1 = B1p.value(iB1);
-        float const cosa = cos(B1 * seq.alpha * M_PI / 180.f);
[…]
