+++ src/sim-mupa.h
@@ -7,10 +7,11 @@
 namespace Sim {
 
 Result MUPA(
-  Parameter const T1,
-  Parameter const T2,
-  Parameter const B1,
-  Sequence const seq,
-  Log &log);
+    Parameter const T1,
+    Parameter const T2,
+    Parameter const B1,
+    Sequence const seq,
+    long const randN,
+    Log &log);
 
 }

[…]
