+++ src/sim-prep.h
@@ -7,10 +7,11 @@
 namespace Sim {
 
 Result Simple(
-  Parameter const T1p,
-  Parameter const betap,
-  Parameter const B1p,
-  Sequence const seq,
-  Log &log);
+    Parameter const T1p,
+    Parameter const betap,
+    Parameter const B1p,
+    Sequence const seq,
+    long const nRand,
+    Log &log);
 
 } // namespace Sim
