+++ src/main_basis_sim.cpp
@@ -70,12 +70,12 @@ int main_basis_sim(args::Subparser &parser)
   Sim::Result result;
   if (ng) {
     Sim::Parameter const gamma{ng.Get(), gLo.Get(), gHi.Get(), false};
-    result = Sim::Eddy(T1, beta, gamma, B1, seq, log);
+    result = Sim::Eddy(T1, beta, gamma, B1, seq, randomSamp.Get(), log);
   } else if (mupa) {
     Sim::Parameter const T2{65, 0.02, 0.2, true};
     result = Sim::MUPA(T1, T2, B1, seq, randomSamp.Get(), log);
   } else {
-    result = Sim::Simple(T1, beta, B1, seq, log);
+    result = Sim::Simple(T1, beta, B1, seq, randomSamp.Get(), log);
   }
 
   // Normalize dictionary
