+++ src/sim-eddy.cpp
@@ -11,22 +11,24 @@ Result Eddy(
     Parameter const gammap,
     Parameter const B1p,
     Sequence const seq,
+    long const nRand,
     Log &log)
 {
   log.info("Eddy Current MP-ZTE simulation");
   log.info(
-      FMT_STRING("SPS {}, FA {}, TR {}s, TI {}s, Trec {}s"),
+      FMT_STRING("SPS {}, FA {}, TR {}s, Trec {}s, Tramp {}s, Tssi {}s"),
       seq.sps,
       seq.alpha,
       seq.TR,
-      seq.TI,
-      seq.Trec);
+      seq.Trec,
+      seq.Tramp,
+      seq.Tssi);
   log.info(FMT_STRING("{} values of T1 from {} to {}s"), T1p.N, T1p.lo, T1p.hi);
   log.info(FMT_STRING("{} values of β from {} to {}"), betap.N, betap.lo, betap.hi);
   log.info(FMT_STRING("{} values of ɣ from {} to {}"), gammap.N, gammap.lo, gammap.hi);
   log.info(FMT_STRING("{} values of B1 from {} to {}"), B1p.N, B1p.lo, B1p.hi);
   ParameterGenerator<4> gen({T1p, betap, gammap, B1p});
-  long const totalN = gen.totalN();
+  long totalN = (nRand > 0) ? nRand : gen.totalN();
   Result result;
   result.dynamics.resize(totalN, 4 * seq.sps);
   result.parameters.resize(totalN, 4);
@@ -35,19 +37,19 @@ Result Eddy(
   auto task = [&](long const lo, long const hi, long const ti) {
     for (long ip = lo; ip < hi; ip++) {
       log.progress(ip, lo, hi);
-      auto const P = gen.values(ip);
+      auto const P = (nRand > 0) ? gen.rand() : gen.values(ip);
       // Set up matrices
       float const T1 = P(0);
       float const R1 = 1.f / T1;
       float const e1 = exp(-R1 * seq.TR);
-      Eigen::Matrix2f E1;
-      E1 << e1, 1 - e1, 0.f, 1.f;
-      float const einv = exp(-R1 * seq.TI);
-      Eigen::Matrix2f Ei;
-      Ei << einv, 1 - einv, 0.f, 1.f;
+      float const eramp = exp(-R1 * seq.Tramp);
+      float const essi = exp(-R1 * seq.Tssi);
       float const erec = exp(-R1 * seq.Trec);
-      Eigen::Matrix2f Er;
-      Er << erec, 1 - erec, 0.f, 1.f;
+      Eigen::Matrix2f E1, Eramp, Essi, Einv, Erec;
+      E1 << e1, 1 - e1, 0.f, 1.f;
+      Eramp << eramp, 1 - eramp, 0.f, 1.f;
+      Essi << essi, 1 - essi, 0.f, 1.f;
+      Erec << erec, 1 - erec, 0.f, 1.f;
 
       float const B1 = P(1);
       float const cosa = cos(B1 * seq.alpha * M_PI / 180.f);
@@ -56,47 +58,41 @@ Result Eddy(
       A << cosa, 0.f, 0.f, 1.f;
 
       float const beta = P(3);
-      Eigen::Matrix2f B;
-      B << beta, 0.f, 0.f, 1.f;
-
       float const gamma = P(2);
       Eigen::Matrix2f PC0, PC1, PC2, PC3;
       float pinc = M_PI / 2.f;
-      PC0 << cos(gamma), 0.f, 0.f, 1.f;
[…]
