+++ src/sim-eddy.h
@@ -7,11 +7,12 @@
 namespace Sim {
 
 Result Eddy(
-  Parameter const T1,
-  Parameter const beta,
-  Parameter const gamma,
-  Parameter const B1,
-  Sequence const seq,
-  Log &log);
+    Parameter const T1,
+    Parameter const beta,
+    Parameter const gamma,
+    Parameter const B1,
+    Sequence const seq,
+    long const nRand,
+    Log &log);
 
 }
