### 2021-11-05 670157c
Start refactoring IO
