+++ src/main_basis_sim.cpp
@@ -12,23 +12,23 @@ int main_basis_sim(args::Subparser &parser)
   args::Positional<std::string> oname(parser, "OUTPUT", "Name for the basis file");
 
   args::ValueFlag<long> sps(
-      parser, "SEGMENT LENGTH", "Number of spokes/readouts per segment", {'s', "spokes"}, 128);
+    parser, "SEGMENT LENGTH", "Number of spokes/readouts per segment", {'s', "spokes"}, 128);
   args::ValueFlag<float> alpha(parser, "FLIP ANGLE", "Read-out flip-angle", {'a', "alpha"}, 1.);
   args::ValueFlag<float> TR(parser, "TR", "Read-out repetition time", {"tr"}, 0.002f);
   args::ValueFlag<float> Tramp(parser, "Tramp", "Ramp up/down times", {"tramp"}, 0.01f);
   args::ValueFlag<float> Tssi(parser, "Tssi", "Inter-segment time", {"tssi"}, 0.012f);
   args::ValueFlag<float> TI(
-      parser, "TI", "Inversion time (from prep to segment start)", {"ti"}, 0.45f);
+    parser, "TI", "Inversion time (from prep to segment start)", {"ti"}, 0.45f);
   args::ValueFlag<float> Trec(
-      parser, "TREC", "Recover time (from segment end to prep)", {"trec"}, 0.f);
+    parser, "TREC", "Recover time (from segment end to prep)", {"trec"}, 0.f);
   args::ValueFlag<long> nT1(
-      parser, "N", "Number of T1 values for basis (default 128)", {"nT1"}, 32);
+    parser, "N", "Number of T1 values for basis (default 128)", {"nT1"}, 32);
   args::ValueFlag<float> T1Lo(
-      parser, "T1", "Low value of T1 for basis (default 0.4s)", {"T1lo"}, 0.4f);
+    parser, "T1", "Low value of T1 for basis (default 0.4s)", {"T1lo"}, 0.4f);
   args::ValueFlag<float> T1Hi(
-      parser, "T1", "High value of T1 for basis (default 5s)", {"T1hi"}, 5.f);
+    parser, "T1", "High value of T1 for basis (default 5s)", {"T1hi"}, 5.f);
   args::ValueFlag<long> nb(
-      parser, "N", "Number of beta values for basis (default 128)", {"nbeta"}, 32);
+    parser, "N", "Number of beta values for basis (default 128)", {"nbeta"}, 32);
   args::ValueFlag<float> bLo(parser, "β", "Low value for β (default -1)", {"betalo"}, -1.f);
   args::ValueFlag<float> bHi(parser, "β", "High value for β (default 1)", {"betahi"}, 1.f);
   args::Flag bLog(parser, "", "Use logarithmic spacing for β", {"betalog"});
@@ -39,31 +39,31 @@ int main_basis_sim(args::Subparser &parser)
 
   args::ValueFlag<long> ng(parser, "N", "Number of eddy-current angles", {"eddy"}, 32);
   args::ValueFlag<float> gLo(
-      parser, "ɣ", "Low value for eddy-current angles (default -π)", {"eddylo"}, -M_PI);
+    parser, "ɣ", "Low value for eddy-current angles (default -π)", {"eddylo"}, -M_PI);
   args::ValueFlag<float> gHi(
-      parser, "ɣ", "High value for eddy-current angles (default π)", {"eddyhi"}, M_PI);
+    parser, "ɣ", "High value for eddy-current angles (default π)", {"eddyhi"}, M_PI);
 
   args::Flag mupa(parser, "M", "Run a MUPA simulation", {"mupa"});
 
   args::ValueFlag<long> randomSamp(
[…]
