+++ src/io_hd5.hpp
@@ -36,7 +36,7 @@ hid_t type_impl(type_tag<std::complex<T>>)
   status = H5Tinsert(complex_id, "i", HOFFSET(complex_t, i), scalar_id);
   if (status) {
     throw std::runtime_error(
-        "Exception occurred creating complex datatype " + std::to_string(status));
+      "Exception occurred creating complex datatype " + std::to_string(status));
   }
   return complex_id;
 }
@@ -49,10 +49,10 @@ hid_t type()
 
 template <typename Scalar, int ND>
 void store_tensor(
-    Handle const &parent,
-    std::string const &name,
-    Eigen::Tensor<Scalar, ND> const &data,
-    Log const &log)
+  Handle const &parent,
+  std::string const &name,
+  Eigen::Tensor<Scalar, ND> const &data,
+  Log const &log)
 {
   herr_t status;
   hsize_t ds_dims[ND], chunk_dims[ND];
@@ -73,6 +73,36 @@ void store_tensor(
   status = H5Pset_chunk(plist, ND, chunk_dims);
 
   hid_t const tid = type<Scalar>();
+  hid_t const dset = H5Dcreate(parent, name.c_str(), tid, space, H5P_DEFAULT, plist, H5P_DEFAULT);
+  status = H5Dwrite(dset, tid, H5S_ALL, H5S_ALL, H5P_DEFAULT, data.data());
+  status = H5Pclose(plist);
+  status = H5Sclose(space);
+  status = H5Dclose(dset);
+  if (status) {
+    Log::Fail("Could not write tensor {}, code: {}", name, status);
+  } else {
+    log.info("Wrote dataset: {}", name);
+  }
+}
+
+void store_matrix(
+  Handle const &parent,
+  std::string const &name,
+  Eigen::Ref<Eigen::MatrixXf const> const &data,
+  Log const &log)
+{
+  herr_t status;
+  hsize_t ds_dims[2], chunk_dims[2];
+  // HD5=row-major, Eigen=col-major, so need to reverse the dimensions
+  ds_dims[0] = data.cols();
+  ds_dims[1] = data.rows();
+  std::copy_n(ds_dims, 2, chunk_dims);
+  auto const space = H5Screate_simple(2, ds_dims, NULL);
+  auto const plist = H5Pcreate(H5P_DATASET_CREATE);
+  status = H5Pset_deflate(plist, 2);
+  status = H5Pset_chunk(plist, 2, chunk_dims);
+
+  hid_t const tid = type<float>();
   hid_t const dset = H5Dcreate(parent, name.c_str(), tid, space, H5P_DEFAULT, plist, H5P_DEFAULT);
   status = H5Dwrite(dset, tid, H5S_ALL, H5S_ALL, H5P_DEFAULT, data.data());
   status = H5Pclose(plist);
@@ -107,10 +137,7 @@ Eigen::DSizes<long, ND> get_dims(Handle const &parent, std::string const &name, 
 
 template <typename Scalar, int ND>
 void load_tensor(
-    Handle const &parent,
-    std::string const &name,
-    Eigen::Tensor<Scalar, ND> &tensor,
-    Log const &log)
+  Handle const &parent, std::string const &name, Eigen::Tensor<Scalar, ND> &tensor, Log const &log)
 {
   hid_t dset = H5Dopen(parent, name.c_str(), H5P_DEFAULT);
   if (dset < 0) {
@@ -123,13 +150,13 @@ void load_tensor(
   for (int ii = 0; ii < ND; ii++) {
     if ((long)dims[ii] != tensor.dimension(ii)) {
       Log::Fail(
-          FMT_STRING("Expected dimensions were {}, but were {} on disk"),
[…]

[…]
