### 2021-11-06 5f5e6cc
Removed B1 from prep sim
