+++ src/main_phantom.cpp
@@ -1,7 +1,6 @@
 
 #include "coils.h"
-#include "io_hd5.h"
-#include "io_nifti.h"
+#include "io.h"
 #include "log.h"
 #include "op/recon.h"
 #include "parse_args.h"
@@ -18,43 +17,43 @@ int main_phantom(args::Subparser &parser)
 {
   args::Positional<std::string> iname(parser, "FILE", "Filename to write phantom data to");
   args::ValueFlag<std::string> oftype(
-      parser, "OUT FILETYPE", "File type of output (nii/nii.gz/img/h5)", {"oft"}, "h5");
+    parser, "OUT FILETYPE", "File type of output (nii/nii.gz/img/h5)", {"oft"}, "h5");
   args::ValueFlag<float> osamp(parser, "OSAMP", "Grid oversampling factor (2)", {'s', "os"}, 2.f);
   args::Flag kb(parser, "KB", "Use Kaiser-Bessel interpolation", {"kb"});
   args::ValueFlag<float> fov(
-      parser, "FOV", "Field of View in mm (default 256)", {'f', "fov"}, 240.f);
+    parser, "FOV", "Field of View in mm (default 256)", {'f', "fov"}, 240.f);
   args::ValueFlag<long> matrix(parser, "MATRIX", "Matrix size (default 128)", {'m', "matrix"}, 128);
   args::Flag shepplogan(parser, "SHEPP-LOGAN", "3D Shepp-Logan phantom", {"shepp_logan"});
   args::ValueFlag<float> phan_r(
-      parser, "RADIUS", "Radius of the spherical phantom in mm (default 90)", {"phan_rad"}, 90.f);
+    parser, "RADIUS", "Radius of the spherical phantom in mm (default 90)", {"phan_rad"}, 90.f);
   args::ValueFlag<Eigen::Vector3f, Vector3fReader> phan_c(
-      parser, "X,Y,Z", "Center position of phantom (in mm)", {"center"}, Eigen::Vector3f::Zero());
+    parser, "X,Y,Z", "Center position of phantom (in mm)", {"center"}, Eigen::Vector3f::Zero());
   args::ValueFlag<Eigen::Vector3f, Vector3fReader> phan_rot(
-      parser, "ax,ay,az", "Rotation of phantom (in deg)", {"rotation"}, Eigen::Vector3f::Zero());
+    parser, "ax,ay,az", "Rotation of phantom (in deg)", {"rotation"}, Eigen::Vector3f::Zero());
   args::ValueFlag<long> coil_rings(
-      parser, "COIL RINGS", "Number of rings in coil (default 1)", {"rings"}, 1);
+    parser, "COIL RINGS", "Number of rings in coil (default 1)", {"rings"}, 1);
   args::ValueFlag<float> coil_r(
-      parser, "COIL RADIUS", "Radius of the coil in mm (default 150)", {"coil_rad"}, 150.f);
+    parser, "COIL RADIUS", "Radius of the coil in mm (default 150)", {"coil_rad"}, 150.f);
   args::ValueFlag<float> read_samp(
-      parser, "SRATE", "Read-out oversampling (default 2)", {'r', "read"}, 2);
+    parser, "SRATE", "Read-out oversampling (default 2)", {'r', "read"}, 2);
   args::ValueFlag<float> spoke_samp(
-      parser, "SRATE", "Spoke undersampling (default 1)", {'s', "spokes"}, 1);
+    parser, "SRATE", "Spoke undersampling (default 1)", {'s', "spokes"}, 1);
   args::ValueFlag<long> lores(
-      parser, "LO-RES", "Include lo-res k-space with scale factor (suggest 8)", {'l', "lores"}, 0);
[…]
