+++ src/traj_spirals.h
@@ -3,4 +3,4 @@
 #include "info.h"
 
 R3 ArchimedeanSpiral(Info const &info);
-R3 Phyllotaxis(Info const &info, long const smoothness, long const spi);
+R3 Phyllotaxis(Info const &info, long const smoothness, long const spi, bool const gmeans);
