### 2021-11-09 80f0a4d
Tweaks to SDC & KB Kernel
