### 2021-11-09 4e3b237
Refactored Recon/Grid ops
