+++ src/main_defs.h
@@ -2,23 +2,21 @@
 
 #include "parse_args.h"
 
-int main_basis_admm(args::Subparser &parser);
-int main_basis_blend(args::Subparser &parser);
-int main_basis_cg(args::Subparser &parser);
-int main_basis_dict(args::Subparser &parser);
-int main_basis_phantom(args::Subparser &parser);
-int main_basis_recon(args::Subparser &parser);
-int main_basis_sim(args::Subparser &parser);
+int main_admm(args::Subparser &parser);
+int main_blend(args::Subparser &parser);
+int main_lookup(args::Subparser &parser);
 int main_cg(args::Subparser &parser);
 int main_compress(args::Subparser &parser);
 int main_espirit(args::Subparser &parser);
 int main_grid(args::Subparser &parser);
 int main_hdr(args::Subparser &parser);
 int main_phantom(args::Subparser &parser);
+int main_phantom_basis(args::Subparser &parser);
 int main_plan(args::Subparser &parser);
 int main_recon(args::Subparser &parser);
 int main_sdc(args::Subparser &parser);
 int main_sense(args::Subparser &parser);
+int main_sim(args::Subparser &parser);
 int main_split(args::Subparser &parser);
 int main_tgv(args::Subparser &parser);
 int main_traj(args::Subparser &parser);
