+++ src/op/grid-basis-kb.cpp
@@ -162,7 +162,7 @@ void GridBasisKB<InPlane, ThroughPlane>::A(Input const &cart, Output &noncart) c
       auto const si = mapping_.sortedIndices[ii];
       auto const c = mapping_.cart[si];
       auto const nc = mapping_.noncart[si];
-      auto const k = kernel_(mapping_.offset[si], 1.f / basisScale_);
+      auto const k = kernel_(mapping_.offset[si], basisScale_);
       stC.set(2, c.x - (InPlane / 2));
       stC.set(3, c.y - (InPlane / 2));
       stC.set(4, c.z - (ThroughPlane / 2));
