+++ src/main_admm.cpp
@@ -76,7 +76,7 @@ int main_admm(args::Subparser &parser)
   }
   log.info("All Volumes: {}", log.toNow(all_start));
   WriteBasisVolumes(
-    out, basis, mag, info, iname.Get(), oname.Get(), "basis-admm", oftype.Get(), log);
+    out, basis, mag, info, iname.Get(), oname.Get(), "admm", oftype.Get(), log);
   FFT::End(log);
   return EXIT_SUCCESS;
 }
