### 2021-11-10 00ac1be
Fix CG debugging output
