+++ src/cg.hpp
@@ -36,7 +36,10 @@ void cg(long const &max_its, float const &thresh, Op const &op, typename Op::Inp
     float const beta = r_new / r_old;
     p.device(dev) = r + p * p.constant(beta);
     float const delta = r_new / norm_x0;
-    log.image(x, fmt::format(FMT_STRING("cg-b-{:02}.nii"), icg));
+    log.image(x, fmt::format(FMT_STRING("cg-x-{:02}.nii"), icg));
+    log.image(p, fmt::format(FMT_STRING("cg-p-{:02}.nii"), icg));
+    log.image(q, fmt::format(FMT_STRING("cg-q-{:02}.nii"), icg));
+    log.image(r, fmt::format(FMT_STRING("cg-r-{:02}.nii"), icg));
     log.info(FMT_STRING("CG {}: ɑ {} β {} δ {}"), icg, alpha, beta, delta);
     if (delta < thresh) {
       log.info(FMT_STRING("Reached convergence threshold"));
@@ -48,13 +51,13 @@ void cg(long const &max_its, float const &thresh, Op const &op, typename Op::Inp
 
 template <typename Op>
 void cgvar(
-    long const &max_its,
-    float const &thresh,
-    float const &pre0,
-    float const &pre1,
-    Op &op,
-    typename Op::Input &x,
-    Log &log)
+  long const &max_its,
+  float const &thresh,
+  float const &pre0,
+  float const &pre1,
+  Op &op,
+  typename Op::Input &x,
+  Log &log)
 {
   log.info(FMT_STRING("Starting Variably Preconditioned Conjugate Gradients"));
   auto dev = Threads::GlobalDevice();
