### 2021-11-16 9922209
Document vcpkg dependencies better
