### 2021-11-17 f3ef8f7
Add a kernel test
