### 2021-11-19 c705bf0
Add some kind of diffusion blending
