+++ src/main_defs.h
@@ -4,12 +4,13 @@
 
 int main_admm(args::Subparser &parser);
 int main_blend(args::Subparser &parser);
-int main_lookup(args::Subparser &parser);
 int main_cg(args::Subparser &parser);
 int main_compress(args::Subparser &parser);
 int main_espirit(args::Subparser &parser);
 int main_grid(args::Subparser &parser);
 int main_hdr(args::Subparser &parser);
+int main_lookup(args::Subparser &parser);
+int main_meta(args::Subparser &parser);
 int main_phantom(args::Subparser &parser);
 int main_phantom_basis(args::Subparser &parser);
 int main_plan(args::Subparser &parser);
