+++ src/op/grid-basis-kb.cpp
@@ -121,11 +121,13 @@ void GridBasisKB<InPlane, ThroughPlane>::Adj(Output const &noncart, Input &cart)
     }
   };
 
-  auto const &start = log_.now();
+  auto const start = log_.now();
   cart.setZero();
   Threads::RangeFor(grid_task, mapping_.cart.size());
+  log_.debug("Basis Non-cart -> Cart: {}", log_.toNow(start));
   if (safe_) {
     log_.info("Combining thread workspaces...");
+    auto const start2 = log_.now();
     for (long ti = 0; ti < nThreads; ti++) {
       if (szZ[ti]) {
         cart
@@ -136,8 +138,8 @@ void GridBasisKB<InPlane, ThroughPlane>::Adj(Output const &noncart, Input &cart)
           .device(dev) += workspace[ti];
       }
     }
+    log_.debug("Combining took: {}", log_.toNow(start2));
   }
-  log_.debug("Non-cart -> Cart: {}", log_.toNow(start));
 }
 
 template <int InPlane, int ThroughPlane>
