### 2021-11-22 565e115
Slight speed-up for basis gridding
