+++ bench/grid.cpp
@@ -1,32 +1,35 @@
 #define CATCH_CONFIG_ENABLE_BENCHMARKING
 
+#include "../src/op/grid.h"
 #include "../src/info.h"
+#include "../src/op/grid-basis-kb.h"
+#include "../src/op/grid-basis-nn.h"
 #include "../src/op/grid-kb.h"
 #include "../src/op/grid-nn.h"
-#include "../src/op/grid.h"
 #include "../src/traj_spirals.h"
 
 #include <catch2/catch.hpp>
 
-TEST_CASE("Grid-NN")
+TEST_CASE("Grid")
 {
   Log log;
-  long const M = 64;
+  long const M = 32;
   long const C = 8;
-  Info const info{.type = Info::Type::ThreeD,
-                  .channels = C,
-                  .matrix = Eigen::Array3l::Constant(M),
-                  .read_points = M / 2,
-                  .read_gap = 0,
-                  .spokes_hi = M * M,
-                  .spokes_lo = 0,
-                  .lo_scale = 1.f,
-                  .volumes = 1,
-                  .echoes = 1,
-                  .tr = 1.f,
-                  .voxel_size = Eigen::Array3f::Constant(1.f),
-                  .origin = Eigen::Array3f::Constant(0.f),
-                  .direction = Eigen::Matrix3f::Identity()};
+  Info const info{
+    .type = Info::Type::ThreeD,
+    .channels = C,
+    .matrix = Eigen::Array3l::Constant(M),
+    .read_points = M / 2,
+    .read_gap = 0,
+    .spokes_hi = M * M,
+    .spokes_lo = 0,
+    .lo_scale = 1.f,
+    .volumes = 1,
+    .echoes = 1,
+    .tr = 1.f,
+    .voxel_size = Eigen::Array3f::Constant(1.f),
+    .origin = Eigen::Array3f::Constant(0.f),
+    .direction = Eigen::Matrix3f::Identity()};
   auto const points = ArchimedeanSpiral(info);
   Trajectory traj(info, points, log);
 
@@ -38,53 +41,49 @@ TEST_CASE("Grid-NN")
   };
 
   GridNN gridnn(traj, os, false, log);
+  GridKB<5, 5> gridkb(traj, os, false, log);
+
   auto nc = info.noncartesianVolume();
-  auto c = gridnn.newMultichannel(C);
+  auto c = gridkb.newMultichannel(C);
+  long const nB = 4;
+  R2 basis(256, nB);
+  basis.setConstant(1.f);
+
+  GridBasisNN gridbnn(traj, os, false, basis, log);
+  GridBasisKB<5, 5> gridbkb(traj, os, false, basis, log);
+  Cx5 b(C, nB, c.dimension(1), c.dimension(2), c.dimension(3));
 
   BENCHMARK("NN Noncartesian->Cartesian")
   {
     gridnn.Adj(nc, c);
   };
-  BENCHMARK("NN Cartesian->Noncartesian")
+  BENCHMARK("NN Basis Noncartesian->Cartesian")
   {
-    gridnn.A(c, nc);
+    gridbnn.Adj(nc, b);
   };
-}
-
-TEST_CASE("Grid-KB")
-{
-  Log log;
-  long const M = 64;
-  long const C = 8;
-  Info const info{.type = Info::Type::ThreeD,
-                  .channels = C,
-                  .matrix = Eigen::Array3l::Constant(M),
-                  .read_points = M / 2,
-                  .read_gap = 0,
-                  .spokes_hi = M * M,
-                  .spokes_lo = 0,
-                  .lo_scale = 1.f,
-                  .volumes = 1,
[…]
