### 2021-11-23 72e7e90
Big fixes for Pipe method
