+++ src/op/grid-basis-kernel.hpp
@@ -1,0 +1,201 @@
+#pragma once
+
+#include "../cropper.h"
+#include "../kernel.h"
+#include "../trajectory.h"
+#include "grid-basis.h"
+
+template <typename Kernel>
+struct GridBasis final : GridBasisOp
+{
+  GridBasis(
+    Trajectory const &traj,
+    float const os,
+    bool const unsafe,
+    R2 const &basis,
+    Log &log,
+    float const inRes = -1.f,
+    bool const shrink = false)
+    : GridBasisOp(traj.mapping(os, (Kernel::InPlane / 2), inRes, shrink), unsafe, basis, log)
+    , kernel_{os}
+  {
+  }
+
+  GridBasis(Mapping const &mapping, bool const unsafe, R2 const &basis, Log &log)
+    : GridBasisOp(mapping, unsafe, basis, log)
+    , kernel_{mapping_.osamp}
+  {
+  }
+
+  void A(Input const &cart, Output &noncart) const
+  {
+    assert(cart.dimension(0) == noncart.dimension(0));
+    assert(cart.dimension(1) == basis_.dimension(1));
+    assert(cart.dimension(2) == mapping_.cartDims[0]);
+    assert(cart.dimension(3) == mapping_.cartDims[1]);
+    assert(cart.dimension(4) == mapping_.cartDims[2]);
+
+    long const nchan = cart.dimension(0);
+    long const nB = basis_.dimension(1);
+    Eigen::IndexList<int, int, FixIn, FixIn, FixThrough> szC;
+    szC.set(0, nchan);
+    szC.set(1, nB);
+
+    auto grid_task = [&](long const lo, long const hi) {
+      Eigen::IndexList<FixZero, FixZero, int, int, int> stC;
+      for (auto ii = lo; ii < hi; ii++) {
+        log_.progress(ii, lo, hi);
+        auto const si = mapping_.sortedIndices[ii];
+        auto const c = mapping_.cart[si];
+        auto const n = mapping_.noncart[si];
+        auto const b = (basis_.chip(n.spoke % basis_.dimension(0), 0) * basisScale_).cast<Cx>();
+        auto const k = kernel_(mapping_.offset[si]);
+        stC.set(2, c.x - (Kernel::InPlane / 2));
+        stC.set(3, c.y - (Kernel::InPlane / 2));
+        stC.set(4, c.z - (Kernel::ThroughPlane / 2));
+        noncart.chip(n.spoke, 2).chip(n.read, 1) =
+          cart.slice(stC, szC)
+            .contract(b, Eigen::IndexPairList<Eigen::type2indexpair<1, 0>>())
+            .contract(
+              k.template cast<Cx>(),
+              Eigen::IndexPairList<
+                Eigen::type2indexpair<1, 0>,
+                Eigen::type2indexpair<2, 1>,
+                Eigen::type2indexpair<3, 2>>());
+      }
+    };
+    auto const &start = log_.now();
+    noncart.setZero();
+    Threads::RangeFor(grid_task, mapping_.cart.size());
+    log_.debug("Cart -> Non-cart: {}", log_.toNow(start));
+  }
+
+  void Adj(Output const &noncart, Input &cart) const
+  {
+    assert(cart.dimension(0) == noncart.dimension(0));
+    assert(cart.dimension(1) == basis_.dimension(1));
+    assert(cart.dimension(2) == mapping_.cartDims[0]);
+    assert(cart.dimension(3) == mapping_.cartDims[1]);
+    assert(cart.dimension(4) == mapping_.cartDims[2]);
[…]
