### 2021-11-25 598fa6e
Fix clang cast bug
