+++ src/op/grid-basis-kernel.hpp
@@ -48,7 +48,8 @@ struct GridBasis final : GridBasisOp
         auto const si = mapping_.sortedIndices[ii];
         auto const c = mapping_.cart[si];
         auto const n = mapping_.noncart[si];
-        auto const b = (basis_.chip(n.spoke % basis_.dimension(0), 0) * basisScale_).cast<Cx>();
+        auto const b =
+          (basis_.chip(n.spoke % basis_.dimension(0), 0) * basisScale_).template cast<Cx>();
         auto const k = kernel_(mapping_.offset[si]);
         stC.set(2, c.x - (Kernel::InPlane / 2));
         stC.set(3, c.y - (Kernel::InPlane / 2));
@@ -132,11 +133,12 @@ struct GridBasis final : GridBasisOp
         auto const c = mapping_.cart[si];
         auto const n = mapping_.noncart[si];
         auto const nc = noncart.chip(n.spoke, 2).chip(n.read, 1);
-        auto const b = (basis_.chip(n.spoke % basis_.dimension(0), 0) * basisScale_).cast<Cx>();
-        auto const k = kernel_(mapping_.offset[si]).template cast<Cx>();
+        auto const b = (basis_.chip(n.spoke % basis_.dimension(0), 0) * basisScale_);
+        auto const k = kernel_(mapping_.offset[si]);
         ncb = (nc * nc.constant(mapping_.sdc[si])).reshape(rshNC).broadcast(brdNC) *
-              b.reshape(rshB).broadcast(brdB);
-        auto const nbk = ncb.reshape(rshNCB).broadcast(brdNCB) * k.reshape(rshK).broadcast(brdK);
+              b.template cast<Cx>().reshape(rshB).broadcast(brdB);
+        auto const nbk = ncb.reshape(rshNCB).broadcast(brdNCB) *
+                         k.template cast<Cx>().reshape(rshK).broadcast(brdK);
 
         stC.set(2, c.x - (Kernel::InPlane / 2));
         stC.set(3, c.y - (Kernel::InPlane / 2));
@@ -162,12 +164,11 @@ struct GridBasis final : GridBasisOp
           cart
             .slice(
               Sz5{0, 0, 0, 0, minZ[ti]},
-              Sz5{
-                cart.dimension(0),
-                cart.dimension(1),
-                cart.dimension(2),
-                cart.dimension(3),
-                szZ[ti]})
+              Sz5{cart.dimension(0),
+                  cart.dimension(1),
+                  cart.dimension(2),
+                  cart.dimension(3),
+                  szZ[ti]})
             .device(dev) += workspace[ti];
         }
       }
