### 2021-11-25 aacbaa7
Adds pipenn SDC method
