+++ src/main_tgv.cpp
@@ -35,7 +35,7 @@ int main_tgv(args::Subparser &parser)
   auto const &info = traj.info();
 
   auto gridder = make_grid(traj, osamp.Get(), kernel.Get(), fastgrid, log);
-  R2 const w = SDC::Choose(sdc.Get(), traj, log);
+  R2 const w = SDC::Choose(sdc.Get(), traj, osamp.Get(), log);
   gridder->setSDC(w);
   Cx4 senseMaps = senseFile ? LoadSENSE(senseFile.Get(), log)
                             : DirectSENSE(
