+++ src/main_cg.cpp
@@ -29,7 +29,7 @@ int main_cg(args::Subparser &parser)
   Trajectory const traj = reader.readTrajectory();
   Info const &info = traj.info();
   auto gridder = make_grid(traj, osamp.Get(), kernel.Get(), fastgrid, log);
-  R2 const w = SDC::Choose(sdc.Get(), traj, log);
+  R2 const w = SDC::Choose(sdc.Get(), traj, osamp.Get(), log);
   gridder->setSDC(w);
   Cx4 senseMaps = senseFile ? LoadSENSE(senseFile.Get(), log)
                             : DirectSENSE(
