+++ src/main_espirit.cpp
@@ -35,7 +35,7 @@ int main_espirit(args::Subparser &parser)
   auto const &info = traj.info();
   log.info(FMT_STRING("Cropping data to {} mm effective resolution"), res.Get());
   auto gridder = make_grid(traj, osamp.Get(), kernel.Get(), fastgrid, log, res);
-  gridder->setSDC(SDC::Pipe(traj, log));
+  gridder->setSDC(SDC::Pipe(traj, true, osamp.Get(), log));
   long const totalCalRad = kRad.Get() + calRad.Get() + (info.spokes_lo ? 0 : info.read_gap);
   Cropper cropper(info, gridder->gridDims(), fov.Get(), log);
   Cx4 sense = cropper.crop4(ESPIRIT(
