+++ test/op/grid.cpp
@@ -1,4 +1,6 @@
 #include "../../src/op/grid.h"
+#include "../../src/sdc.h"
+#include "../../src/tensorOps.h"
 #include "../../src/traj_spirals.h"
 #include "../../src/trajectory.h"
 
@@ -6,43 +8,48 @@
 
 TEST_CASE("ops-grid", "[ops]")
 {
-  //   long const M = 16;
-  //   Info const info{.type = Info::Type::ThreeD,
-  //                   .channels = nchan.Get(),
-  //                   .matrix = Eigen::Array3l::Constant(M),
-  //                   .read_points = M / 2,
-  //                   .read_gap = 0,
-  //                   .spokes_hi = M * M,
-  //                   .spokes_lo = 0,
-  //                   .lo_scale = 1.f,
-  //                   .volumes = 1,
-  //                   .echoes = 1,
-  //                   .tr = 1.f,
-  //                   .voxel_size = Eigen::Array3f::Constant(1.f),
-  //                   .origin = Eigen::Array3f::Constant(0.f),
-  //                   .direction = Eigen::Matrix3f::Identity()};
-  //   points = ArchimedeanSpiral(info);
+  Log log;
+  long const M = 16;
+  float const os = 2.f;
+  Info const info{
+    .type = Info::Type::ThreeD,
+    .channels = 1,
+    .matrix = Eigen::Array3l::Constant(M),
+    .read_points = long(os * M / 2),
+    .read_gap = 0,
+    .spokes_hi = long(M * M),
+    .spokes_lo = 0,
+    .lo_scale = 1.f,
+    .volumes = 1,
+    .echoes = 1,
+    .tr = 1.f,
+    .voxel_size = Eigen::Array3f::Constant(1.f),
+    .origin = Eigen::Array3f::Constant(0.f),
+    .direction = Eigen::Matrix3f::Identity()};
+  auto const points = ArchimedeanSpiral(info);
+  Trajectory const traj(info, points, log);
+  R2 const sdc = SDC::Pipe(traj, true, os, log);
 
-  //   // With credit to PyLops
-  //   SECTION("Dot Test")
-  //   {
-  //     Cx3 x(mapSz, mapSz, mapSz), u(mapSz, mapSz, mapSz);
-  //     Cx4 maps(channels, mapSz, mapSz, mapSz), y(channels, gridSz, gridSz, gridSz),
-  //         v(channels, gridSz, gridSz, gridSz);
+  /* I don't think the classic Dot test from PyLops is applicable to gridding,
+   * because it would not be correct to have a random radial k-space. The k0
+   * samples should all be the same, not random. Hence instead I calculate
+   * y = Adj*A*x for NN and then check if Dot(x,y) is the same as Dot(y,y).
+   * Can't check for y = x because samples not on the radial spokes will be
+   * missing, and checking for KB kernel is not valid because the kernel blurs
+   * the grid.
+   */
+  SECTION("NN-Dot")
+  {
+    auto grid = make_grid(traj, os, Kernels::NN, false, log);
+    grid->setSDC(sdc);
+    Cx4 x = grid->newMultichannel(1), y = grid->newMultichannel(1);
+    Cx3 r(info.channels, info.read_points, info.spokes_total());
 
-  //     v.setRandom();
-  //     u.setRandom();
-  //     // The maps need to be normalized for the Dot test
-  //     maps.setRandom();
-  //     Cx3 rss = ConjugateSum(maps, maps).sqrt();
-  //     maps = maps / Tile(rss, channels);
-
-  //     SenseOp sense(maps, y.dimensions());
[…]
