+++ src/op/recon-basis.h
@@ -3,27 +3,27 @@
 #include "operator.h"
 
 #include "../fft_plan.h"
-#include "grid-basis.h"
+#include "gridBase.h"
 #include "sense.h"
 
 struct ReconBasisOp final : Operator<4, 3>
 {
-  ReconBasisOp(GridBasisOp *gridder, Cx4 const &maps, Log &log);
+  ReconBasisOp(GridBase *gridder, Cx4 const &maps, Log &log);
 
   void A(Input const &x, Output &y) const;
   void Adj(Output const &x, Input &y) const;
-  // void AdjA(Input const &x, Input &y) const;
+  void AdjA(Input const &x, Input &y) const;
 
   Sz3 dimensions() const;
   Sz3 outputDimensions() const;
   void calcToeplitz(Info const &info);
 
 private:
-  GridBasisOp *gridder_;
+  GridBase *gridder_;
   Cx5 mutable grid_;
   Cx5 transfer_;
-  SenseBasisOp sense_;
+  SenseOp sense_;
   R3 apo_;
-  FFT::ThreeDBasis fft_;
+  FFT::Planned<5, 3> fft_;
   Log log_;
 };
