+++ src/main.cpp
@@ -15,7 +15,6 @@ int main(int const argc, char const *const argv[])
   args::Command lookup(commands, "lookup", "Basis dictionary lookup", &main_lookup);
   args::Command meta(commands, "meta", "Print meta-data entries", &main_meta);
   args::Command phantom(commands, "phantom", "Construct a digitial phantom", &main_phantom);
-  args::Command phantom_basis(commands, "phantom-basis", "Phantom with basis", &main_phantom_basis);
   args::Command plan(commands, "plan", "Plan FFTs", &main_plan);
   args::Command recon(commands, "recon", "Basic non-iterative reconstruction", &main_recon);
   args::Command sdc(commands, "sdc", "Calculate Sample Density Compensation", &main_sdc);
@@ -23,7 +22,7 @@ int main(int const argc, char const *const argv[])
   args::Command sim(commands, "sim", "Simulate a basis set", &main_sim);
   args::Command split(commands, "split", "Split data", &main_split);
   args::Command traj(commands, "traj", "Write out the trajectory and PSF", &main_traj);
-  args::Command tgv(commands, "tgv", "Iterative TGV regularised recon", &main_tgv);
+  // args::Command tgv(commands, "tgv", "Iterative TGV regularised recon", &main_tgv);
   args::Command version(commands, "version", "Print version number", &main_version);
   args::Command zinfandel(commands, "zinfandel", "ZINFANDEL k-space filling", &main_zinfandel);
   args::GlobalOptions globals(parser, global_group);
