+++ src/op/gridBase.h
@@ -13,7 +13,6 @@ struct GridBase : Operator<5, 3>
 
   virtual R3 apodization(Sz3 const sz) const = 0; // Calculate the apodization factor for this grid
 
-  Sz3 gridDims() const; // Returns the dimensions of the grid
   void setSDC(float const dc);
   void setSDC(R2 const &sdc);
   void setUnsafe();
