+++ src/tgv.cpp
@@ -3,133 +3,125 @@
 #include "tensorOps.h"
 #include "threads.h"
 
-inline auto ForwardDiff(Cx3 const &a, Eigen::Index const d)
+inline auto ForwardDiff(Cx4 const &a, Eigen::Index const d)
 {
-  Sz3 const sz{a.dimension(0) - 2, a.dimension(1) - 2, a.dimension(2) - 2};
-  Sz3 const st1{1, 1, 1};
-  Sz3 fwd{1, 1, 1};
-  fwd[d] = 2;
+  Sz4 const sz{a.dimension(0), a.dimension(1) - 2, a.dimension(2) - 2, a.dimension(3) - 2};
+  Sz4 const st1{0, 1, 1, 1};
+  Sz4 fwd{0, 1, 1, 1};
+  fwd[d + 1] = 2;
 
   return (a.slice(fwd, sz) - a.slice(st1, sz));
 }
 
-inline auto BackwardDiff(Cx3 const &a, Eigen::Index const d)
+inline auto BackwardDiff(Cx4 const &a, Eigen::Index const d)
 {
-  Sz3 const sz{a.dimension(0) - 2, a.dimension(1) - 2, a.dimension(2) - 2};
-  Sz3 const st1{1, 1, 1};
-  Sz3 bck{1, 1, 1};
-  bck[d] = 0;
+  Sz4 const sz{a.dimension(0), a.dimension(1) - 2, a.dimension(2) - 2, a.dimension(3) - 2};
+  Sz4 const st1{0, 1, 1, 1};
+  Sz4 bck{0, 1, 1, 1};
+  bck[d + 1] = 0;
 
   return (a.slice(st1, sz) - a.slice(bck, sz));
 }
 
-inline auto CentralDiff(Cx3 const &a, Eigen::Index const d)
+inline auto CentralDiff(Cx4 const &a, Eigen::Index const d)
 {
-  Sz3 const sz{a.dimension(0) - 2, a.dimension(1) - 2, a.dimension(2) - 2};
-  Sz3 const st1{1, 1, 1};
-  Sz3 fwd{1, 1, 1};
-  Sz3 bck{1, 1, 1};
-  fwd[d] = 2;
-  bck[d] = 0;
+  Sz4 const sz{a.dimension(0), a.dimension(1) - 2, a.dimension(2) - 2, a.dimension(3) - 2};
+  Sz4 const st1{0, 1, 1, 1};
+  Sz4 fwd{0, 1, 1, 1};
+  Sz4 bck{0, 1, 1, 1};
+  fwd[d + 1] = 2;
+  bck[d + 1] = 0;
 
   return (a.slice(fwd, sz) - a.slice(bck, sz)) / a.slice(st1, sz).constant(2.f);
 }
 
-inline void Grad(Cx3 const &a, Cx4 &g, Eigen::ThreadPoolDevice &dev)
+inline void Grad(Cx4 const &a, Cx5 &g, Eigen::ThreadPoolDevice &dev)
 {
-  Sz3 const sz{a.dimension(0) - 2, a.dimension(1) - 2, a.dimension(2) - 2};
-  Sz3 const st1{1, 1, 1};
-  g.chip<3>(0).slice(st1, sz).device(dev) = ForwardDiff(a, 0);
-  g.chip<3>(1).slice(st1, sz).device(dev) = ForwardDiff(a, 1);
-  g.chip<3>(2).slice(st1, sz).device(dev) = ForwardDiff(a, 2);
+  Sz4 const sz{a.dimension(0), a.dimension(1) - 2, a.dimension(2) - 2, a.dimension(3) - 2};
+  Sz4 const st1{0, 1, 1, 1};
[…]
