### 2021-11-30 2d79813
Add checks for invalid echoes
