+++ src/main_phantom.cpp
@@ -80,21 +80,20 @@ int main_phantom(args::Subparser &parser)
     auto const spokes_hi =
       sps.Get() *
       ((std::lrint(nex.Get() * matrix.Get() * matrix.Get()) + sps.Get() - 1) / sps.Get());
-    info = Info{
-      .type = Info::Type::ThreeD,
-      .channels = nchan.Get(),
-      .matrix = Eigen::Array3l::Constant(matrix.Get()),
-      .read_points = (long)read_samp.Get() * matrix.Get() / 2,
-      .read_gap = 0,
-      .spokes_hi = spokes_hi,
-      .spokes_lo = 0,
-      .lo_scale = lores ? lores.Get() : 1.f,
-      .volumes = 1,
-      .echoes = 1,
-      .tr = 1.f,
-      .voxel_size = Eigen::Array3f::Constant(fov.Get() / matrix.Get()),
-      .origin = Eigen::Array3f::Constant(-fov.Get() / 2.f),
-      .direction = Eigen::Matrix3f::Identity()};
+    info = Info{.type = Info::Type::ThreeD,
+                .channels = nchan.Get(),
+                .matrix = Eigen::Array3l::Constant(matrix.Get()),
+                .read_points = (long)read_samp.Get() * matrix.Get() / 2,
+                .read_gap = 0,
+                .spokes_hi = spokes_hi,
+                .spokes_lo = 0,
+                .lo_scale = lores ? lores.Get() : 1.f,
+                .volumes = 1,
+                .echoes = 1,
+                .tr = 1.f,
+                .voxel_size = Eigen::Array3f::Constant(fov.Get() / matrix.Get()),
+                .origin = Eigen::Array3f::Constant(-fov.Get() / 2.f),
+                .direction = Eigen::Matrix3f::Identity()};
     if (phyllo) {
       points = Phyllotaxis(info, smoothness.Get(), sps.Get() * spi.Get(), gmeans);
     } else {
@@ -143,7 +142,7 @@ int main_phantom(args::Subparser &parser)
       phan.dimension(0));
   }
   for (long ii = 0; ii < phan.dimension(0); ii++) {
-    phan.chip(ii, 0) =
+    phan.chip<0>(ii) =
       shepplogan
         ? SheppLoganPhantom(
             info.matrix,
@@ -179,21 +178,20 @@ int main_phantom(args::Subparser &parser)
       // GridOp does funky stuff to merge k-spaces. Sample lo-res as if it was hi-res
       lowres_scale = lores.Get();
       auto const spokes_lo = info.spokes_hi / lowres_scale;
-      lo_info = Info{
-        .type = Info::Type::ThreeD,
-        .channels = info.channels,
-        .matrix = info.matrix,
-        .read_points = info.read_points,
-        .read_gap = 0,
-        .spokes_hi = spokes_lo,
-        .spokes_lo = 0,
-        .lo_scale = 1.f,
-        .volumes = 1,
-        .echoes = 1,
-        .tr = 1.f,
-        .voxel_size = info.voxel_size,
-        .origin = info.origin,
-        .direction = Eigen::Matrix3f::Identity()};
+      lo_info = Info{.type = Info::Type::ThreeD,
+                     .channels = info.channels,
+                     .matrix = info.matrix,
+                     .read_points = info.read_points,
+                     .read_gap = 0,
+                     .spokes_hi = spokes_lo,
+                     .spokes_lo = 0,
+                     .lo_scale = 1.f,
+                     .volumes = 1,
+                     .echoes = 1,
+                     .tr = 1.f,
+                     .voxel_size = info.voxel_size,
+                     .origin = info.origin,
[…]
