+++ src/main_admm.cpp
@@ -76,7 +76,7 @@ int main_admm(args::Subparser &parser)
     recon.Adj(reader.noncartesian(iv), vol); // Initialize
     admm(admm_its.Get(), lsq_its.Get(), thr.Get(), recon, reg_rho.Get(), reg, vol, log);
     cropped = out_cropper.crop4(vol);
-    out.chip(iv, 4) = cropped;
+    out.chip<4>(iv) = cropped;
     log.info("Volume {}: {}", iv, log.toNow(vol_start));
   }
   log.info("All Volumes: {}", log.toNow(all_start));
