+++ src/op/grid-basis-kernel.hpp
@@ -49,12 +49,12 @@ struct GridBasis final : GridBasisOp
         auto const c = mapping_.cart[si];
         auto const n = mapping_.noncart[si];
         auto const b =
-          (basis_.chip(n.spoke % basis_.dimension(0), 0) * basisScale_).template cast<Cx>();
+          (basis_.chip<0>(n.spoke % basis_.dimension(0)) * basisScale_).template cast<Cx>();
         auto const k = kernel_(mapping_.offset[si]);
         stC.set(2, c.x - (Kernel::InPlane / 2));
         stC.set(3, c.y - (Kernel::InPlane / 2));
         stC.set(4, c.z - (Kernel::ThroughPlane / 2));
-        noncart.chip(n.spoke, 2).chip(n.read, 1) =
+        noncart.template chip<2>(n.spoke).template chip<1>(n.read) =
           cart.slice(stC, szC)
             .contract(b, Eigen::IndexPairList<Eigen::type2indexpair<1, 0>>())
             .contract(
@@ -132,8 +132,8 @@ struct GridBasis final : GridBasisOp
         auto const si = mapping_.sortedIndices[ii];
         auto const c = mapping_.cart[si];
         auto const n = mapping_.noncart[si];
-        auto const nc = noncart.chip(n.spoke, 2).chip(n.read, 1);
-        auto const b = (basis_.chip(n.spoke % basis_.dimension(0), 0) * basisScale_);
+        auto const nc = noncart.template chip<2>(n.spoke).template chip<1>(n.read);
+        auto const b = (basis_.chip<0>(n.spoke % basis_.dimension(0)) * basisScale_);
         auto const k = kernel_(mapping_.offset[si]);
         ncb = (nc * nc.constant(mapping_.sdc[si])).reshape(rshNC).broadcast(brdNC) *
               b.template cast<Cx>().reshape(rshB).broadcast(brdB);
@@ -164,12 +164,11 @@ struct GridBasis final : GridBasisOp
           cart
             .slice(
               Sz5{0, 0, 0, 0, minZ[ti]},
-              Sz5{
-                cart.dimension(0),
-                cart.dimension(1),
-                cart.dimension(2),
-                cart.dimension(3),
-                szZ[ti]})
+              Sz5{cart.dimension(0),
+                  cart.dimension(1),
+                  cart.dimension(2),
+                  cart.dimension(3),
+                  szZ[ti]})
             .device(dev) += workspace[ti];
         }
       }

[…]
