### 2021-11-30 b647fa3
Removes writeOutput
