+++ src/op/grid-kernel.hpp
@@ -34,11 +34,11 @@ struct Grid final : GridOp
     assert(cart.dimension(3) == mapping_.cartDims[1]);
     assert(cart.dimension(4) == mapping_.cartDims[2]);
 
-    long const nchan = cart.dimension(0);
+    Index const nchan = cart.dimension(0);
     Eigen::IndexList<int, FixIn, FixIn, FixThrough> szC;
     szC.set(0, nchan);
 
-    auto grid_task = [&](long const lo, long const hi) {
+    auto grid_task = [&](Index const lo, Index const hi) {
       Eigen::IndexList<FixZero, int, int, int> stC;
       for (auto ii = lo; ii < hi; ii++) {
         log_.progress(ii, lo, hi);
@@ -73,7 +73,7 @@ struct Grid final : GridOp
     assert(cart.dimension(4) == mapping_.cartDims[2]);
     assert(mapping_.sortedIndices.size() == mapping_.cart.size());
 
-    long const nchan = cart.dimension(0);
+    Index const nchan = cart.dimension(0);
     Eigen::IndexList<int, FixOne, FixOne, FixOne> rshNC;
     constexpr Eigen::IndexList<FixOne, FixIn, FixIn, FixThrough> brdNC;
     rshNC.set(0, nchan);
@@ -86,16 +86,16 @@ struct Grid final : GridOp
     szC.set(0, nchan);
 
     auto dev = Threads::GlobalDevice();
-    long const nThreads = dev.numThreads();
+    Index const nThreads = dev.numThreads();
     std::vector<Cx5> workspace(nThreads);
-    std::vector<long> minZ(nThreads, 0L), szZ(nThreads, 0L);
-    auto grid_task = [&](long const lo, long const hi, long const ti) {
+    std::vector<Index> minZ(nThreads, 0L), szZ(nThreads, 0L);
+    auto grid_task = [&](Index const lo, Index const hi, Index const ti) {
       // Allocate working space for this thread
       Eigen::IndexList<FixZero, int, int, int> stC;
       minZ[ti] = mapping_.cart[mapping_.sortedIndices[lo]].z - ((Kernel::ThroughPlane - 1) / 2);
 
       if (safe_) {
-        long const maxZ =
+        Index const maxZ =
           mapping_.cart[mapping_.sortedIndices[hi - 1]].z + (Kernel::ThroughPlane / 2);
         szZ[ti] = maxZ - minZ[ti] + 1;
         workspace[ti].resize(
@@ -132,7 +132,7 @@ struct Grid final : GridOp
     if (safe_) {
       log_.info("Combining thread workspaces...");
       auto const start2 = log_.now();
-      for (long ti = 0; ti < nThreads; ti++) {
+      for (Index ti = 0; ti < nThreads; ti++) {
         if (szZ[ti]) {
           cart
             .slice(
@@ -160,7 +160,7 @@ struct Grid final : GridOp
     fft.reverse(temp);
     R3 a = Crop3(R3(temp.real()), sz);
     float const scale =
-      sqrt(std::accumulate(gridSz.cbegin(), gridSz.cend(), 1, std::multiplies<long>()));
+      sqrt(std::accumulate(gridSz.cbegin(), gridSz.cend(), 1, std::multiplies<Index>()));
     log_.info(
       FMT_STRING("Apodization size {} scale factor: {}"), fmt::join(a.dimensions(), ","), scale);
     a.device(Threads::GlobalDevice()) = a * a.constant(scale);
