### 2021-11-30 1520274
Improve end-of-spoke check
