### 2021-11-30 b358ecc
Start converting to .readTensor
