### 2021-12-01 ad97266
Fix sundry bugs
