+++ src/op/recon.cpp
@@ -77,7 +77,7 @@ void ReconOp::Adj(Output const &x, Input &y) const
 void ReconOp::AdjA(Input const &x, Input &y) const
 {
   if (transfer_.size() == 0) {
-    Output temp;
+    Output temp(gridder_->mapping().noncartDims);
     A(x, temp);
     Adj(temp, y);
   } else {
