### 2021-12-01 3a18c0d
Make Töplitz an explicit argument
