### 2021-12-01 ec6cf83
Add sdcPow back as an option
