+++ src/main_sdc.cpp
@@ -13,7 +13,6 @@ int main_sdc(args::Subparser &parser)
     parser, "SDC", "SDC type: 'pipe', 'pipenn', 'radial'", {"sdc"}, "pipe");
   args::ValueFlag<std::string> oftype(
     parser, "OUT FILETYPE", "File type of output (nii/nii.gz/img/h5)", {"oft"}, "h5");
-  args::ValueFlag<float> sdcPow(parser, "P", "SDC Power (default 1.0)", {'p', "pow"}, 1.0f);
   args::ValueFlag<float> osamp(parser, "OS", "Oversampling when using pipenn", {'s', "os"}, 2.f);
   Log log = ParseCommand(parser, iname);
   HD5::Reader reader(iname.Get(), log);
@@ -29,9 +28,6 @@ int main_sdc(args::Subparser &parser)
     dc = SDC::Radial(traj, log);
   } else {
     Log::Fail(FMT_STRING("Uknown SDC method: {}"), sdc.Get());
-  }
-  if (sdcPow) {
-    dc = dc.pow(sdcPow.Get());
   }
   HD5::Writer writer(OutName(iname.Get(), oname.Get(), "sdc", "h5"), log);
   writer.writeInfo(info);
