+++ src/main_plan.cpp
@@ -21,17 +21,14 @@ int main_plan(args::Subparser &parser)
   HD5::Reader reader(iname.Get(), log);
   auto const traj = reader.readTrajectory();
   auto gridder = make_grid(traj, osamp.Get(), kernel.Get(), fastgrid, log);
-  Cx5 grid4(gridder->inputDimensions(traj.info().channels, 1));
-  Cx5 grid3(gridder->inputDimensions(1, 1));
-  FFT::Planned<5, 3> fft3(grid3, log);
-  FFT::Planned<5, 3> fft4(grid4, log);
+  FFT::Planned<5, 3> fft3(gridder->inputDimensions(traj.info().channels), log);
+  FFT::Planned<5, 3> fft4(gridder->inputDimensions(1), log);
 
   if (basisFile) {
     HD5::Reader basisReader(basisFile.Get(), log);
     R2 basis = basisReader.readBasis();
     auto gridderBasis = make_grid_basis(traj, osamp.Get(), kernel.Get(), fastgrid, basis, log);
-    Cx5 grid5(gridderBasis->inputDimensions(traj.info().channels));
-    FFT::Planned<5, 3> fft(grid5, log);
+    FFT::Planned<5, 3> fft(gridderBasis->inputDimensions(traj.info().channels), log);
   }
 
   FFT::End(log);
