### 2021-12-02 22e449c
Add a benchmark for kernels
