### 2021-12-05 a7415b2
Move apodization into SENSE for ReconOp
