+++ src/op/sense.cpp
@@ -1,9 +1,10 @@
+#include "sense.h"
 #include "../tensorOps.h"
 #include "../threads.h"
-#include "sense.h"
 
-SenseOp::SenseOp(Cx4 const &maps, Output::Dimensions const &bigSize)
+SenseOp::SenseOp(Cx4 const &maps, Output::Dimensions const &bigSize, Log &log)
   : maps_{maps}
+  , log_{log}
 {
   size_[0] = maps_.dimension(0);
   size_[1] = bigSize[1];
@@ -40,6 +41,16 @@ SenseOp::OutputDims SenseOp::outputDimensions() const
   return full_;
 }
 
+void SenseOp::setApodization(GridBase *gridder)
+{
+  apo_ = gridder->apodization(Sz3{size_[2], size_[3], size_[4]});
+}
+
+void SenseOp::resetApodization()
+{
+  apo_ = R3();
+}
+
 void SenseOp::A(Input const &x, Output &y) const
 {
   assert(x.dimension(1) == maps_.dimension(1));
@@ -50,7 +61,7 @@ void SenseOp::A(Input const &x, Output &y) const
   assert(y.dimension(2) == (maps_.dimension(1) + left_[2] + right_[2]));
   assert(y.dimension(3) == (maps_.dimension(2) + left_[3] + right_[3]));
   assert(y.dimension(4) == (maps_.dimension(3) + left_[4] + right_[4]));
-
+  log_.info("SENSE Forward");
   Eigen::IndexList<FixOne, int, int, int, int> resX;
   resX.set(1, x.dimension(0));
   resX.set(2, x.dimension(1));
@@ -73,8 +84,21 @@ void SenseOp::A(Input const &x, Output &y) const
       return std::make_pair(left, right);
     });
 
-  y.device(Threads::GlobalDevice()) =
-    (x.reshape(resX).broadcast(brdX) * maps_.reshape(resMaps).broadcast(brdMaps)).pad(paddings);
+  if (apo_.size()) {
+    Eigen::IndexList<FixOne, int, int, int> resApo;
+    resApo.set(1, apo_.dimension(0));
+    resApo.set(2, apo_.dimension(1));
+    resApo.set(3, apo_.dimension(2));
+    Eigen::IndexList<int, FixOne, FixOne, FixOne> brdApo;
+    brdApo.set(0, x.dimension(0));
+    y.device(Threads::GlobalDevice()) =
+      ((x / apo_.cast<Cx>().reshape(resApo).broadcast(brdApo)).reshape(resX).broadcast(brdX) *
+       maps_.reshape(resMaps).broadcast(brdMaps))
+        .pad(paddings);
+  } else {
+    y.device(Threads::GlobalDevice()) =
+      (x.reshape(resX).broadcast(brdX) * maps_.reshape(resMaps).broadcast(brdMaps)).pad(paddings);
+  }
 }
 
 void SenseOp::Adj(Output const &x, Input &y) const
@@ -87,7 +111,7 @@ void SenseOp::Adj(Output const &x, Input &y) const
   assert(y.dimension(1) == maps_.dimension(1));
   assert(y.dimension(2) == maps_.dimension(2));
   assert(y.dimension(3) == maps_.dimension(3));
-
+  log_.info("SENSE Reverse");
   Eigen::IndexList<int, FixOne, int, int, int> resMaps;
   resMaps.set(0, maps_.dimension(0));
   resMaps.set(2, maps_.dimension(1));
@@ -96,8 +120,20 @@ void SenseOp::Adj(Output const &x, Input &y) const
   Eigen::IndexList<FixOne, int, FixOne, FixOne, FixOne> brdMaps;
[…]
