+++ src/op/recon.cpp
@@ -7,11 +7,11 @@
 ReconOp::ReconOp(GridBase *gridder, Cx4 const &maps, Log &log)
   : gridder_{gridder}
   , grid_{gridder_->inputDimensions(maps.dimension(0))}
-  , sense_{maps, grid_.dimensions()}
-  , apo_{gridder_->apodization(Sz3{maps.dimension(1), maps.dimension(2), maps.dimension(3)})}
+  , sense_{maps, grid_.dimensions(), log}
   , fft_{grid_, log}
   , log_{log}
 {
+  sense_.setApodization(gridder);
 }
 
 ReconOp::InputDims ReconOp::inputDimensions() const
@@ -32,24 +32,13 @@ void ReconOp::calcToeplitz(Info const &info)
   Cx3 tf(1, info.read_points, info.spokes_total());
   gridder_->A(transfer_, tf);
   gridder_->Adj(tf, transfer_);
+  sense_.resetApodization();
 }
 
 void ReconOp::A(Input const &x, Output &y) const
 {
-  auto dev = Threads::GlobalDevice();
   auto const &start = log_.now();
-
-  Index const nB = x.dimension(0);
-  Eigen::IndexList<FixOne, int, int, int> rshA;
-  Eigen::IndexList<int, FixOne, FixOne, FixOne> brdA;
-  rshA.set(1, apo_.dimension(0));
-  rshA.set(2, apo_.dimension(1));
-  rshA.set(3, apo_.dimension(2));
-  brdA.set(0, nB);
-
-  Cx4 apodised(x.dimensions());
-  apodised.device(dev) = x / apo_.cast<Cx>().reshape(rshA).broadcast(brdA);
-  sense_.A(apodised, grid_);
+  sense_.A(x, grid_); // SENSE takes care of apodization
   fft_.forward(grid_);
   gridder_->A(grid_, y);
   log_.debug("Encode: {}", log_.toNow(start));
@@ -57,20 +46,10 @@ void ReconOp::A(Input const &x, Output &y) const
 
 void ReconOp::Adj(Output const &x, Input &y) const
 {
-  auto dev = Threads::GlobalDevice();
   auto const &start = log_.now();
   gridder_->Adj(x, grid_);
   fft_.reverse(grid_);
-  sense_.Adj(grid_, y);
-
-  Index const nB = y.dimension(0);
-  Eigen::IndexList<FixOne, int, int, int> rshA;
-  Eigen::IndexList<int, FixOne, FixOne, FixOne> brdA;
-  rshA.set(1, apo_.dimension(0));
-  rshA.set(2, apo_.dimension(1));
-  rshA.set(3, apo_.dimension(2));
-  brdA.set(0, nB);
-  y.device(dev) = y / apo_.cast<Cx>().reshape(rshA).broadcast(brdA);
+  sense_.Adj(grid_, y); // SENSE takes care of apodization
   log_.debug("Decode: {}", log_.toNow(start));
 }
 
