### 2021-12-05 087bea2
Fixes loading points when echoes exists
