+++ src/main_defs.h
@@ -11,6 +11,7 @@ int main_grid(args::Subparser &parser);
 int main_hdr(args::Subparser &parser);
 int main_lookup(args::Subparser &parser);
 int main_meta(args::Subparser &parser);
+int main_nii(args::Subparser &parser);
 int main_phantom(args::Subparser &parser);
 int main_phantom_basis(args::Subparser &parser);
 int main_plan(args::Subparser &parser);
