+++ src/main_split.cpp
@@ -8,6 +8,7 @@ int main_split(args::Subparser &parser)
 {
   args::Positional<std::string> iname(parser, "FILE", "HD5 file to recon");
   args::ValueFlag<std::string> oname(parser, "OUTPUT", "Override output name", {'o', "out"});
+  args::ValueFlag<Index> lores(parser, "N", "Extract first N spokes as lo-res", {'l', "lores"});
   args::ValueFlag<Index> nspokes(parser, "SPOKES", "Spokes per segment", {"n", "nspokes"});
   args::ValueFlag<Index> vol(parser, "VOLUME", "Only take this volume", {"v", "vol"}, 0);
   args::ValueFlag<float> ds(parser, "DS", "Downsample by factor", {"ds"}, 1.0);
@@ -19,55 +20,65 @@ int main_split(args::Subparser &parser)
   auto const traj = reader.readTrajectory();
   auto info = traj.info();
   info.volumes = 1; // Only output one volume
-  R3 all_points = traj.points();
-  Cx3 all_ks = reader.noncartesian(vol.Get());
+  R3 points = traj.points();
+  Cx3 ks = reader.noncartesian(vol.Get());
 
-  if (info.spokes_lo) {
+  if (ds) {
+    info.read_points = (Index)std::round((float)info.read_points / ds.Get());
+    float const scalef = (float)info.read_points / (float)info.read_points;
+    info.voxel_size = info.voxel_size * scalef;
+    info.matrix = (info.matrix.cast<float>() / scalef).cast<Index>();
+    points = points * scalef;
+    ks = ks.slice(Sz3{0, 0, 0}, Sz3{info.channels, info.read_points, info.spokes_total()});
+  }
+
+  if (lores) {
     Info lo_info = info;
     lo_info.spokes_lo = 0;
-    lo_info.spokes_hi = info.spokes_lo;
+    lo_info.spokes_hi = lores.Get();
     lo_info.lo_scale = 0.f;
     R3 lo_points =
-      all_points.slice(Sz3{0, 0, 0}, Sz3{3, info.read_points, info.spokes_lo}) / info.lo_scale;
-    Cx4 lo_ks = all_ks.slice(Sz3{0, 0, 0}, Sz3{info.channels, info.read_points, info.spokes_lo})
-                  .reshape(Sz4{info.channels, info.read_points, info.spokes_lo, 1});
+      points.slice(Sz3{0, 0, 0}, Sz3{3, info.read_points, lores.Get()}) / info.lo_scale;
+    Cx4 lo_ks = ks.slice(Sz3{0, 0, 0}, Sz3{info.channels, info.read_points, lores.Get()})
+                  .reshape(Sz4{info.channels, info.read_points, lores.Get(), 1});
     HD5::Writer writer(OutName(iname.Get(), oname.Get(), "lores", "h5"), log);
     writer.writeTrajectory(Trajectory(lo_info, lo_points, log));
     writer.writeNoncartesian(lo_ks);
+    info.spokes_lo = 0;
+    info.spokes_hi = info.spokes_total() - lores.Get();
+    info.lo_scale = 0.f;
+    points = points.slice(Sz3{0, lores.Get(), 0}, Sz3{3, info.read_points, info.spokes_hi});
+    ks = ks.slice(Sz3{0, lores.Get(), 0}, Sz3{info.channels, info.read_points, info.spokes_hi});
   }
 
-  Info hi_info = info;
-  hi_info.spokes_lo = 0;
-  hi_info.lo_scale = 0.f;
-  float scalef = 1.0f;
-  if (ds) {
-    hi_info.read_points = (Index)std::round((float)info.read_points / ds.Get());
[…]
