### 2021-12-06 1a2e044
Fixes writing echoes out
