+++ src/trajectory.h
@@ -34,8 +34,9 @@ struct Trajectory
   Trajectory(Info const &info, R3 const &points, Log const &log);
   Trajectory(Info const &info, R3 const &points, I1 const &echoes, Log const &log);
   Info const &info() const;
-
   R3 const &points() const;
+  I1 const &echoes() const;
+
   Point3 point(int16_t const read, int32_t const spoke, float const nomRad) const;
 
   // Return the appropriate filter value for merging lo/hi-res k-spaces
