### 2021-12-06 ed2f025
Fixes check for NaN trajectory points
