### 2021-12-07 106baed
Refactor to only .spokes
