+++ bench/grid.cpp
@@ -17,16 +17,14 @@ TEST_CASE("Grid")
                   .matrix = Eigen::Array3l::Constant(M),
                   .read_points = M / 2,
                   .read_gap = 0,
-                  .spokes_hi = M * M,
-                  .spokes_lo = 0,
-                  .lo_scale = 1.f,
+                  .spokes = M * M,
                   .volumes = 1,
                   .echoes = 1,
                   .tr = 1.f,
                   .voxel_size = Eigen::Array3f::Constant(1.f),
                   .origin = Eigen::Array3f::Constant(0.f),
                   .direction = Eigen::Matrix3f::Identity()};
-  auto const points = ArchimedeanSpiral(info);
+  auto const points = ArchimedeanSpiral(info.read_points, info.spokes);
   Trajectory traj(info, points, log);
 
   float const os = 2.f;

[…]
