### 2021-12-07 c07770a
Fix phantom lo-res & checking info struct
