### 2021-12-13 98dd265
Remove read_gap from info
