### 2021-12-13 5055836
Adds NUFFT command
