+++ src/op/pad.h
@@ -1,9 +1,10 @@
 #pragma once
 
+#include "gridBase.h"
 #include "operator.h"
 
 template <int Rank>
-struct CropOp final : Operator<Rank, Rank>
+struct PadOp final : Operator<Rank, Rank>
 {
   using Parent = Operator<Rank, Rank>;
   using Input = typename Parent::Input;
@@ -11,17 +12,18 @@ struct CropOp final : Operator<Rank, Rank>
   using Output = typename Parent::Output;
   using OutputDims = typename Parent::OutputDims;
 
-  CropOp(InputDims const &bigSize, OutputDims const &smallSize);
-  InputDims bigDimensions() const;
+  PadOp(InputDims const &bigSize, OutputDims const &smallSize);
+  InputDims inputDimensions() const;
   OutputDims outputDimensions() const;
+  void setApodization(GridBase *gridder); // Combine apodization into SENSE to save memory
+  void resetApodization();
 
   void A(Input const &x, Output &y) const;
   void Adj(Output const &x, Input &y) const;
   void AdjA(Input const &x, Input &y) const;
 
 private:
-  InputDims full_, left_, size_, right_;
+  InputDims input_, output_, left_, right_, resApo_, brdApo_;
+  R3 apo_;
+  Eigen::array<std::pair<Index, Index>, Rank> paddings_;
 };
-
-using CropOp3 = CropOp<3>;
-using CropOp4 = CropOp<4>;
