+++ src/padder.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "types.h"
-
-// Embeds a small tensor into a larger one
-void ZeroPad(Cx4 const &small, Cx4 &large);

[…]
