+++ src/sdc.h
@@ -7,5 +7,6 @@
 
 namespace SDC {
 R2 Pipe(Trajectory const &traj, bool const nn, float const os, Log &log);
+R2 Radial(Trajectory const &traj, Index const lores, Index const gap, Log &log);
 R2 Choose(std::string const &fname, Trajectory const &traj, float const os, Log &log);
 } // namespace SDC
