+++ src/main_sdc.cpp
@@ -11,9 +11,10 @@ int main_sdc(args::Subparser &parser)
   args::ValueFlag<std::string> oname(parser, "OUTPUT", "Override output name", {'o', "out"});
   args::ValueFlag<std::string> sdc(
     parser, "SDC", "SDC type: 'pipe', 'pipenn', 'radial'", {"sdc"}, "pipe");
-  args::ValueFlag<std::string> oftype(
-    parser, "OUT FILETYPE", "File type of output (nii/nii.gz/img/h5)", {"oft"}, "h5");
   args::ValueFlag<float> osamp(parser, "OS", "Oversampling when using pipenn", {'s', "os"}, 2.f);
+  args::ValueFlag<Index> lores(
+    parser, "L", "Number of lo-res spokes for radial", {'l', "lores"}, 0);
+  args::ValueFlag<Index> gap(parser, "G", "Read-gap for radial", {'g', "gap"}, 0);
   Log log = ParseCommand(parser, iname);
   HD5::Reader reader(iname.Get(), log);
   auto const traj = reader.readTrajectory();
@@ -24,6 +25,8 @@ int main_sdc(args::Subparser &parser)
     dc = SDC::Pipe(traj, false, 2.1f, log);
   } else if (sdc.Get() == "pipenn") {
     dc = SDC::Pipe(traj, true, osamp.Get(), log);
+  } else if (sdc.Get() == "radial") {
+    dc = SDC::Radial(traj, lores.Get(), gap.Get(), log);
   } else {
     Log::Fail(FMT_STRING("Uknown SDC method: {}"), sdc.Get());
   }
