+++ src/io_hd5.cpp
@@ -76,7 +76,7 @@ void CheckInfoType(hid_t handle)
   // Re-orderd fields are okay. Missing is not
   for (auto const &check_name : names) {
     bool found = false;
-    for (int ii = 0; ii < names.size(); ii++) {
+    for (int ii = 0; ii < n_members; ii++) {
       std::string const member_name(H5Tget_member_name(dtype, ii));
       if (member_name == check_name) {
         found = true;
