### 2021-12-13 2ebf9ca
Add echo intensity compensation
