+++ src/trajectory.h
@@ -27,6 +27,7 @@ struct Mapping
   Sz3 cartDims, noncartDims;
   float osamp;
   int8_t echoes;
+  Eigen::ArrayXf echoWeights;
 };
 
 struct Trajectory
@@ -36,6 +37,7 @@ struct Trajectory
   Info const &info() const;
   R3 const &points() const;
   I1 const &echoes() const;
+
   Point3 point(int16_t const read, int32_t const spoke, float const nomRad) const;
 
   // Generate an appropriate mapping
