+++ docs/recon.rst
@@ -12,13 +12,13 @@ By default ``recon`` will output complex-valued images in ``.h5`` format that ha
 
 Non-cartesian MRI data requires a Non-Uniform FFT (NUFFT) instead of a simple FFT for conversion between k-space and image space. The NUFFT consists of a gridding step and then a normal FFT [1, 2]. The gridding step in RIESLING is controlled by the oversampling and kernel options.
 
-``--os=X`` controls the grid oversampling ratio. The default value is 2, but this is very memory intensive (as it is a 3D grid, 2x oversampling requires 8 times the memory of the native matrix size). Reducing the over-sampling to a value of 1.3 leads to a reduced memory foot-print for little impact in image quality [3]. 
+``--os=X`` controls the grid oversampling ratio. The default value is 2, but this is very memory intensive (as it is a 3D grid, 2x oversampling requires 8 times the memory of the native matrix size). Reducing the over-sampling to a value of 1.3 leads to a reduced memory foot-print for little impact in image quality [3]. However, oversampling factors below 2 do not work with Töplitz embedding.
 
-``--kb`` selects a Kaiser-Bessel (KB) kernel instead of the default nearest-neighbour (NN)kernel [3, 4]. Kaiser-Bessel is the default in most other toolboxes and may become the default in RIESLING in the future. The ``--kw`` option controls the kernel width, the default value is 3 and is sufficient for most applications.
+``--kernel=KB3`` selects a width 3 Kaiser-Bessel (KB) kernel instead of the default nearest-neighbour (NN)kernel [3, 4]. Kaiser-Bessel is the default in most other toolboxes and may become the default in RIESLING in the future. KB5 is also available.
 
 The final gridding option is ``--fast-grid``. RIESLING uses a parallelized gridding operation. In order to avoid multiple threads writing to the same Cartesian k-space point, the trajectory is sorted by Cartesian k-space location and each thread processes its own chunk of the sorted co-ordinates. However, there may still be race conditions at the edge of each chunk, particularly for small images with highly oversampled k-space centers. To prevent this, each thread uses its own workspace, and then these are combined into the final grid at the end. This process is thread-safe but doubles the memory requirements for the gridding operations. The ``--fast-grid`` option makes the threads write directly into the final grid, reducing memory consumption but at the risk of conflicting writes into the grid. When gridding high resolution images on a small number of threads, e.g. 10 or fewer, the probability of a race condition is vanishingly small. Use at your own risk.
 
-The gridding step can also compensate for the increased density of samples in the oversampled central k-space region with most non-Cartesian trajectories. This step is often omitted in 2D non-Cartesian iterative reconstructions, but is essential for reasonable convergence in 3D non-Cartesian reconstruction. The ``--sdc`` option controls the sample density compensation method - valid values are "none" to turn it off, "radial" for analytic radial densities, and "pipe" for the iterative method of Pipe et al [5]. The default is "pipe". It is also possible to pre-calculate the densities of a given trajectory using the ``riesling sdc`` command and then pass in the resulting file, i.e. ``--sdc=file-sdc.h5``. The weighting of the compensation can be reduced using the ``--sdc_exp`` option - the exponent is applied to all k-space densities equally and values between 0 and 1 make sense.
[…]

[…]
