### 2021-12-15 bea1eda
Move ADMM Töplitz to option
