### 2021-12-16 1d3fb67
Print wisdom path
