### 2021-12-16 74621bf
Fix cropping bugs
