### 2021-12-17 011ef89
Avoid tensor aliasing
