### 2021-12-18 a58d57f
fix downsampling calculation
