+++ src/main_split.cpp
@@ -26,11 +26,12 @@ int main_split(args::Subparser &parser)
   auto const volName = fmt::format("vol-{:02d}", vol.Get());
 
   if (ds) {
-    info.read_points = (Index)std::round((float)info.read_points / ds.Get());
-    float const scalef = (float)info.read_points / (float)info.read_points;
+    int ds_read_points = (Index)std::round((float)info.read_points / ds.Get());
+    float const scalef = (float)info.read_points / (float)ds_read_points;
     info.voxel_size = info.voxel_size * scalef;
     info.matrix = (info.matrix.cast<float>() / scalef).cast<Index>();
     points = points * scalef;
+    info.read_points = static_cast<Index>(ds_read_points);
     Cx3 ds_ks = ks.slice(Sz3{0, 0, 0}, Sz3{info.channels, info.read_points, info.spokes});
     ks = ds_ks;
   }
@@ -48,9 +49,6 @@ int main_split(args::Subparser &parser)
     writer.writeTrajectory(Trajectory(lo_info, lo_points, lo_echoes, log));
     writer.writeNoncartesian(lo_ks);
     info.spokes -= lores.Get();
-    log.info(FMT_STRING("Reduced spokes. {}, {}"), lores.Get(), info.spokes);
-    log.info(FMT_STRING("points: {}"), points.dimensions());
-
     R3 hi_points = points.slice(Sz3{0, 0, lores.Get()}, Sz3{3, info.read_points, info.spokes});
     Cx3 hi_ks = ks.slice(Sz3{0, 0, lores.Get()}, Sz3{info.channels, info.read_points, info.spokes});
     I1 hi_echoes = echoes.slice(Sz1{lores.Get()}, Sz1{info.spokes});
@@ -66,8 +64,6 @@ int main_split(args::Subparser &parser)
     int const num_int = static_cast<int>((num_full_int - 1) * ns * 1.f / spoke_step + 1);
     log.info(
       FMT_STRING("Interleaves: {} Spokes per interleave: {} Step: {}"), num_int, ns, spoke_step);
-    log.info(FMT_STRING("info.spokes={}"), info.spokes);
-
     int rem_spokes = info.spokes - num_full_int * ns;
     if (rem_spokes > 0) {
       log.info(FMT_STRING("Warning! Last interleave will have {} extra spokes."), rem_spokes);
