### 2022-01-06 37d7c6e
Adds rank checks for tensor reads
