### 2022-01-07 f78cb32
Header-only fmt to avoid visibility problems
