### 2022-01-07 e1111fb
Add distinct trim/blank options
