### 2022-01-07 eba6e70
Explicit gap
